Clean up utils; add CZE league scrape for seasons; Change Kral to CZE league instead of CZE2
