+++ season_scraper/index.js
@@ -2,30 +2,10 @@ const dotenv = require('dotenv')
 const utils = require('../utils')
 const leagueScrapers = require('./leagues')
 
-const LEAGUE_CODES = {
-  OHL: 'ohlScraper',
-  WHL: 'whlScraper',
-  QMJHL: 'qmjhlScraper',
-  AHL: 'ahlScraper',
-  USHL: 'ushlScraper',
-  NLA: 'nlaScraper',
-  KHL: 'khlScraper',
-  ECHL: 'echlScraper',
-  SHL: 'shlScraper',
-  VHL: 'vhlScraper',
-  NCAA: 'ncaaScraper',
-  Liiga: 'liigaScraper',
-  Mestis: 'mestisScraper',
-  CZE2: 'czech2Scraper',
-  Allsv: 'allsvenskanScraper',
-  Sarja20: 'sarja20Scraper',
-  MHL: 'mhlScraper',
-}
-
 dotenv.config()
 
 module.exports = async prospect => {
-  const scraperName = LEAGUE_CODES[prospect.league]
+  const scraperName = leagueScrapers.leagueCodes[prospect.league]
   if (!scraperName) {
     throw new Error(`No scraper found for league: ${prospect.league}`)
   }
