+++ season_scraper/leagues/__tests__/__fixtures__/cze_kral_profile.fixture.js
@@ -1,0 +1,3604 @@
+module.exports = `
+<body class="brand-not-loaded"><div class="fucking-eu-cookies fucking-priority"><span id="eu-text">Tento web používá k poskytování služeb, personalizaci reklam a analýze návštěvnosti soubory cookie. Používáním tohoto webu s tím souhlasíte. </span><a href="https://www.google.com/intl/cs/policies/technologies/cookies/" target="_blank" id="eu-url">Více informací</a><button id="eu-button">Souhlasím</button></div>
+		<script type="text/javascript" src="https://img.cncenter.cz/static/data/blesk/cncenter/sas/js/sas.min.js?v924"></script>
+		<style>
+			.ads{text-align:center;display:block;}
+			.ads-in{display:inline-block;}
+			.ads-in.adFull::before {content: "reklama";font-family:arial;font-size:10px;color:#000;display:block;height:11px;text-align:right;}
+		</style>
+
+	    <div id="fb-root"></div>
+	    <script>(function(d, s, id) {
+	      var js, fjs = d.getElementsByTagName(s)[0];
+	      if (d.getElementById(id)) return;
+	      js = d.createElement(s); js.id = id;
+	      js.src = "//connect.facebook.net/cs_CZ/sdk.js#xfbml=1&version=v2.4&appId=160419347389933";
+	      fjs.parentNode.insertBefore(js, fjs);
+	    }(document, 'script', 'facebook-jssdk'));</script>
+
+    			
+
+		<div class=" banner-page-top">
+            <div class="container">
+		    	<div id="scrolling-div">
+		    		<h3>Hokej.cz na Facebooku <i id="js-close-scrolling-div" class="fa fa-times pull-right"></i></h3>
+		    		<div class="fb-page" data-href="https://www.facebook.com/hokejcz" data-small-header="true" data-adapt-container-width="true" data-hide-cover="false" data-show-facepile="false" data-show-posts="false"><div class="fb-xfbml-parse-ignore"><blockquote cite="https://www.facebook.com/hokejcz"><a href="https://www.facebook.com/hokejcz">Hokej</a></blockquote></div></div>
+		    	</div>
+            </div>
+        </div>
+
+		<div class="sticky-nav">
+
+			<div class="header">
+				<a href="/" class="logo">
+					<img src="/images/logo/logo.svg" title="HOKEJ.CZ" alt="Logo Hokej.cz">
+				</a>
+
+
+<div class="nav-main">
+    <ul class="nav-main__level1">
+        <li>
+			<a href="/tipsport-extraliga?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" title="Tipsport ELH">Tipsport ELH</a>
+
+<div class="nav-main__sub">
+		<div class="nav-main__level2-container">
+			<div class="container">
+				<div class="row">
+					<div class="col-100">
+						<ul class="nav-main__level2">
+							<li><a href="/tipsport-extraliga">Zpravodajství</a></li>
+							<li>
+								<a href="/tipsport-extraliga/tymy">Týmy <i class="fa fa-chevron-down"></i></a>
+<div class="nav-main__sub">
+	<div>
+		<div class="col-100">
+			<div class="row nav-main__teams-container">
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/1721">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Ftlmon-ceb.png&amp;w=76">
+						<span>Č. Budějovice</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/12">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fqe3zb-hc-pardubice-logo-TV-2020-01.png&amp;w=76">
+						<span>Dynamo</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/1711">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fnvl2g-hkr.png&amp;w=76">
+						<span>Hr. Králové</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/5">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Frqvj5-kva.png&amp;w=76">
+						<span>Karlovy Vary</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/22">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fd303u-kom.png&amp;w=76">
+						<span>Kometa</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/18">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fwp8pk-lib.png&amp;w=76">
+						<span>Liberec</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/823">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fg7b2d-lit.png&amp;w=76">
+						<span>Litvínov</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/89">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fmzy6x-mbl.png&amp;w=76">
+						<span>Ml. Boleslav</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/11">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fkngqm-tri.png&amp;w=76">
+						<span>Oceláři</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/24">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2F2dsvo-olo.png&amp;w=76">
+						<span>Olomouc</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/917">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Flvjmh-plz.png&amp;w=76">
+						<span>Plzeň</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/8">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2F3l7xq-spa.png&amp;w=76">
+						<span>Sparta</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/7">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2F2wo6s-vit.png&amp;w=76">
+						<span>Vítkovice</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/562">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fkt1dy-zln.png&amp;w=76">
+						<span>Zlín</span>
+					</a>
+				</div>
+			</div>
+		</div>
+	</div>
+</div>
+							</li>
+							<li><a href="/tipsport-extraliga/zapasy">Zápasy</a></li>
+							<li><a href="/tipsport-extraliga/table">Tabulky</a></li>
+							<li><a href="/tipsport-extraliga/stats-center">Centrum statistik</a></li>
+							<li><a href="/tipsport-extraliga/radegast-index">Radegast Index</a></li>
+							<li><a href="/tipsport-extraliga/visualization">Vizualizace</a></li>
+<li><a href="/zlatahelma">Zlatá helma</a></li>
+<li><a href="/tipsport-extraliga/stranka/21">Partneři</a></li>
+<li><a href="/archiv-clanku/rubrika/312">Hokej v číslech</a></li>
+<li><a href="/tipsport-extraliga/nas-lidr">Náš lídr</a></li>
+						</ul>
+					</div>
+				</div>
+			</div>
+		</div>
+	<div class="container">
+		<div class="row">
+			
+				<div class="col-25">
+			
+			<div>
+		<h2>Tabulka Tipsport extraliga</h2>
+		<table>
+			<tbody>
+				<tr>
+					<td>1.</td>
+					<td><a href="/klub/917">HC Škoda Plzeň</a></td>
+					<td>21</td>
+				</tr>
+				<tr>
+					<td>2.</td>
+					<td><a href="/klub/8">HC Sparta Praha</a></td>
+					<td>18</td>
+				</tr>
+				<tr>
+					<td>3.</td>
+					<td><a href="/klub/1711">Mountfield HK</a></td>
+					<td>17</td>
+				</tr>
+				<tr>
+					<td>4.</td>
+					<td><a href="/klub/11">HC Oceláři Třinec</a></td>
+					<td>14</td>
+				</tr>
+				<tr>
+					<td>5.</td>
+					<td><a href="/klub/18">Bílí Tygři Liberec</a></td>
+					<td>13</td>
+				</tr>
+				<tr>
+					<td>6.</td>
+					<td><a href="/klub/5">HC Energie Karlovy Vary</a></td>
+					<td>12</td>
+				</tr>
+				<tr>
+					<td>7.</td>
+					<td><a href="/klub/89">BK Mladá Boleslav</a></td>
+					<td>11</td>
+				</tr>
+				<tr>
+					<td>8.</td>
+					<td><a href="/klub/7">HC VÍTKOVICE RIDERA</a></td>
+					<td>6</td>
+				</tr>
+				<tr>
+					<td>9.</td>
+					<td><a href="/klub/562">PSG Berani Zlín</a></td>
+					<td>6</td>
+				</tr>
+				<tr>
+					<td>10.</td>
+					<td><a href="/klub/24">HC Olomouc</a></td>
+					<td>6</td>
+				</tr>
+				<tr>
+					<td>11.</td>
+					<td><a href="/klub/22">HC Kometa Brno</a></td>
+					<td>3</td>
+				</tr>
+				<tr>
+					<td>12.</td>
+					<td><a href="/klub/823">HC VERVA Litvínov</a></td>
+					<td>3</td>
+				</tr>
+				<tr>
+					<td>13.</td>
+					<td><a href="/klub/1721">Madeta Motor České Budějovice</a></td>
+					<td>3</td>
+				</tr>
+				<tr>
+					<td>14.</td>
+					<td><a href="/klub/12">HC Dynamo Pardubice</a></td>
+					<td>2</td>
+				</tr>
+			</tbody>
+		</table>
+		<div class="btn-container">
+			<a href="/tipsport-extraliga/table" class="btn btn-black">Podrobná tabulka</a>
+		</div>
+</div>
+
+				</div>
+					<div class="col-25">
+						<h2 class="m-b-10">Kanadské bodování</h2>
+<div class="row">
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/7951">Milan Gulaš</a>
+				<span>11 (5+6)</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/7953">Tomáš Mertl</a>
+				<span>10 (8+2)</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/11974">Matěj Stránský</a>
+				<span>10 (6+4)</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/11296">Petr Straka</a>
+				<span>10 (4+6)</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/6360">Martin Růžička</a>
+				<span>10 (3+7)</span>
+			</p>
+		</div>
+	</div>
+</div>
+						<div class="btn-container">
+							<a href="/tipsport-extraliga/player-stats/detailni" class="btn btn-black">Kompletní statistiky</a>
+						</div>
+					</div>
+					<div class="col-25">
+						<h2 class="m-b-10">Radegast index</h2>
+<div class="row">
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/5075">Roman Polák</a>
+				<span>27</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/9447">Matěj Stříteský</a>
+				<span>27</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/11296">Petr Straka</a>
+				<span>26</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/17410">Vojtěch Budík</a>
+				<span>26</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/16935">Filip Hronek</a>
+				<span>25</span>
+			</p>
+		</div>
+	</div>
+</div>
+						 <div class="btn-container">
+							 <a href="/tipsport-extraliga/player-stats/detailni?section=radegast" class="btn btn-black">Kompletní pořadí RI</a>
+						 </div>
+					</div>
+				<div class="col-25">
+					<div class="banner banner-210" style="padding-top:40px"><a href="/hokejka-clanek/i-vy-muzete-prostrednictvim-hokejka-tv-pomoct-extraligovym-klubum/5050763"><img src="/files/images/84/210x300px-telh.jpg" alt="Hokejka TV" width="210" height="300"></a></div>
+				</div>
+		</div>
+	</div>
+</div>
+		</li>
+        <li>
+			<a href="/reprezentace?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" title="Reprezentace">Reprezentace</a>
+	<div class="nav-main__sub">
+		<div class="nav-main__level2-container">
+			<div class="container">
+				<div class="row">
+					<div class="col-100">
+						<ul class="nav-main__level2">
+<li><a href="/reprezentace/">Zpravodajství</a></li>
+<li><a href="/reprezentace/stranka/27/section">Přehled akcí</a></li>
+<li><a href="/reprezentace/stranka/31/section">Realizační tým</a></li>
+<li><a href="/reprezentace/player-stats/detailni/30?stats-filter-team=28">Statistiky</a></li>
+<li><a href="/reprezentace/stranka/33/section">Starty hráčů</a></li>
+<li><a href="/reprezentace/tournament-archive/15">Archiv MS</a></li>
+<li><a href="/reprezentace/stranka/2500171/section">Partneři</a></li>
+<li><a href="https://fanshop.hokej.cz">fanshop.hokej.cz</a></li>
+<li><a href="/spolecne-tvorime">Společně tvoříme</a></li>
+</ul>
+					</div>
+				</div>
+			</div>
+		</div>
+		<div class="container">
+			<div class="row">
+				<div class="col-25">
+					<div><h2>Seniorské turnaje</h2>
+
+<a href="/reprezentace/zapasy/23"><img src="/files/images/1/thumbnail/karjala-cup-logo2020.png" alt="MS 2020" style="margin-top:10px;  margin-bottom:10px; " width="120" height="92"></a>
+
+</div>
+				</div>
+				<div class="col-25">
+					<div><h2>Juniorské turnaje</h2>
+<!-- <ul class="p-l-5">
+<li><a href="/mladez/zapasy/32">Turnaj pěti U18 v Kravařích</a></li>
+<li><a href="/mladez/zapasy/31">Turnaj čtyř U20 v Hodoníně </a></li>
+</ul> -->
+<a href="/mladez/liga/31"><img src="/files/images/45/ms-u20-2021-small.png" alt="MS do 20 let 2020" style="margin-top:10px; margin-bottom:5px;" width="210" height="106"></a>
+</div>
+				</div>
+				<div class="col-25">
+					<h2>Reprezentační zápasy</h2>
+<ul class="p-l-5">
+<!-- <li><a href="/reprezentace/zapasy/25">Přípravné zápasy</a></li> -->
+<li><a href="/reprezentace/zapasy/23">Karjala Cup 2020</a></li>
+<li><a href="/reprezentace/zapasy/22">Channel One Cup 2019</a></li>
+<li><a href="/reprezentace/zapasy/24">Beijer Hockey Games 2020</a></li>
+<li><a href="/reprezentace/zapasy/48">Euro Hockey Challenge</a></li>
+<li><a href="/reprezentace/zapasy/21">Carlson Hockey Games</a></li>
+<li><a href="/reprezentace/stranka/2554601/section">Euro Hockey Tour</a></li>
+<li><a href="/reprezentace/zapasy/15">Mistrovství světa</a></li>
+</ul>
+				</div>
+				<div class="col-25">
+					<div class="banner banner-210" style="padding-top:40px">
+<a href="https://fanshop.hokej.cz/?utm_source=hokej&amp;utm_medium=banner&amp;utm_campaign=podmenu"><img src="https://hokej.cz/files/images/237/210x260.jpg" alt="nový Fanshop" width="210" height="260"></a></div>
+<!-- <a href="http://goo.gl/DfkBGx"><img src="https://www.hokej.cz/files/images/65/fanshop15-210x300.png" width="210" height="300" alt="Fanshop"></a></div> -->
+				</div>
+			</div>
+		</div>
+	</div>
+		</li>
+        <li>
+			<a href="/chance-liga?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" title="CHANCE LIGA">CHANCE LIGA</a>
+
+<div class="nav-main__sub">
+		<div class="nav-main__level2-container">
+			<div class="container">
+				<div class="row">
+					<div class="col-100">
+						<ul class="nav-main__level2">
+							<li><a href="/chance-liga">Zpravodajství</a></li>
+							<li>
+								<a href="/chance-liga/tymy">Týmy <i class="fa fa-chevron-down"></i></a>
+<div class="nav-main__sub">
+	<div>
+		<div class="col-100">
+			<div class="row nav-main__teams-container">
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/91">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fben.png&amp;w=76">
+						<span>Benátky n. Jiz.</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/132">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fvwnhj-frm.png&amp;w=76">
+						<span>Frýdek-Místek</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/714">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fhav.png&amp;w=76">
+						<span>Havířov</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/16">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2FDUKLA-01-A.png&amp;w=76">
+						<span>Jihlava</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/39">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fkda-logo1-1-.png&amp;w=76">
+						<span>Kadaň</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/828">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fp3bhp-kla.png&amp;w=76">
+						<span>Kladno</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/117">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fkol.png&amp;w=76">
+						<span>Kolín</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/500">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fltm.png&amp;w=76">
+						<span>Litoměřice</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/1927">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fporuba.png&amp;w=76">
+						<span>Poruba</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/66">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fpro.png&amp;w=76">
+						<span>Prostějov</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/128">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fpre.png&amp;w=76">
+						<span>Přerov</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/9">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fwjpve-SLA.png&amp;w=76">
+						<span>Slavia</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/123">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fsok.png&amp;w=76">
+						<span>Sokolov</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/563">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Flogo-drak-hlava-notext-PNG.png&amp;w=76">
+						<span>Šumperk</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/17">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Ftre.png&amp;w=76">
+						<span>Třebíč</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/2611">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2F65uis-unl.png&amp;w=76">
+						<span>Ústí nad Labem</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/1701">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fvrchlabi-logo-2018.png&amp;w=76">
+						<span>Vrchlabí</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/549">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fvsetin.png&amp;w=76">
+						<span>Vsetín</span>
+					</a>
+				</div>
+			</div>
+		</div>
+	</div>
+</div>
+							</li>
+							<li><a href="/chance-liga/zapasy">Zápasy</a></li>
+							<li><a href="/chance-liga/table">Tabulky</a></li>
+							<li><a href="/chance-liga/stats-center">Centrum statistik</a></li>
+						</ul>
+					</div>
+				</div>
+			</div>
+		</div>
+	<div class="container">
+		<div class="row">
+						
+			<div class="col-25">
+				
+				<div>
+		<h2>Tabulka CHANCE LIGA</h2>
+		<table>
+			<tbody>
+				<tr>
+					<td>1.</td>
+					<td><a href="/klub/1927">HC RT TORAX Poruba 2011</a></td>
+					<td>16</td>
+				</tr>
+				<tr>
+					<td>2.</td>
+					<td><a href="/klub/16">HC Dukla Jihlava</a></td>
+					<td>16</td>
+				</tr>
+				<tr>
+					<td>3.</td>
+					<td><a href="/klub/9">HC Slavia Praha</a></td>
+					<td>16</td>
+				</tr>
+				<tr>
+					<td>4.</td>
+					<td><a href="/klub/91">HC Benátky nad Jizerou</a></td>
+					<td>16</td>
+				</tr>
+				<tr>
+					<td>5.</td>
+					<td><a href="/klub/123">HC Baník Sokolov</a></td>
+					<td>15</td>
+				</tr>
+				<tr>
+					<td>6.</td>
+					<td><a href="/klub/828">Rytíři Kladno</a></td>
+					<td>15</td>
+				</tr>
+				<tr>
+					<td>7.</td>
+					<td><a href="/klub/549">VHK ROBE Vsetín</a></td>
+					<td>14</td>
+				</tr>
+				<tr>
+					<td>8.</td>
+					<td><a href="/klub/66">LHK Jestřábi Prostějov</a></td>
+					<td>13</td>
+				</tr>
+				<tr>
+					<td>9.</td>
+					<td><a href="/klub/1701">HC Stadion Vrchlabí</a></td>
+					<td>10</td>
+				</tr>
+				<tr>
+					<td>10.</td>
+					<td><a href="/klub/128">HC ZUBR Přerov</a></td>
+					<td>10</td>
+				</tr>
+				<tr>
+					<td>11.</td>
+					<td><a href="/klub/2611">Hokej Ústí nad Labem s.r.o.</a></td>
+					<td>10</td>
+				</tr>
+				<tr>
+					<td>12.</td>
+					<td><a href="/klub/132">HC Frýdek-Místek</a></td>
+					<td>9</td>
+				</tr>
+				<tr>
+					<td>13.</td>
+					<td><a href="/klub/500">HC Stadion Litoměřice</a></td>
+					<td>8</td>
+				</tr>
+				<tr>
+					<td>14.</td>
+					<td><a href="/klub/563">DRACI PARS ŠUMPERK</a></td>
+					<td>7</td>
+				</tr>
+				<tr>
+					<td>15.</td>
+					<td><a href="/klub/117">SC Kolín</a></td>
+					<td>6</td>
+				</tr>
+				<tr>
+					<td>16.</td>
+					<td><a href="/klub/714">AZ Heimstaden Havířov</a></td>
+					<td>5</td>
+				</tr>
+				<tr>
+					<td>17.</td>
+					<td><a href="/klub/39">SK Trhači Kadaň</a></td>
+					<td>5</td>
+				</tr>
+				<tr>
+					<td>18.</td>
+					<td><a href="/klub/17">SK Horácká Slavia Třebíč</a></td>
+					<td>1</td>
+				</tr>
+			</tbody>
+		</table>
+		<div class="btn-container">
+			<a href="/chance-liga/table" class="btn btn-black">Podrobná tabulka</a>
+		</div>
+</div>
+
+			</div>
+					<div class="col-25">
+						<h2 class="m-b-10">Kanadské bodování</h2>
+<div class="row">
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/31051">Petr Chlán</a>
+				<span>12 (7+5)</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/9427">Ondrej Šedivý</a>
+				<span>11 (6+5)</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/11228">Lukáš Žálčík</a>
+				<span>11 (6+5)</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/12314">David Bartoš</a>
+				<span>11 (3+8)</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/12487">Jaromír Kverka</a>
+				<span>10 (6+4)</span>
+			</p>
+		</div>
+	</div>
+</div>
+						<div class="btn-container">
+							<a href="/chance-liga/player-stats/detailni" class="btn btn-black">Kompletní statistiky</a>
+						</div>
+					</div>
+					<div class="col-25">
+						<h2 class="m-b-10">Tabulka střelců</h2>
+<div class="row">
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/31051">Petr Chlán</a>
+				<span>7</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/9427">Ondrej Šedivý</a>
+				<span>6</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/11228">Lukáš Žálčík</a>
+				<span>6</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/12487">Jaromír Kverka</a>
+				<span>6</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/14940">Radek Veselý</a>
+				<span>6</span>
+			</p>
+		</div>
+	</div>
+</div>
+						 <div class="btn-container">
+							 <a href="/chance-liga/player-stats/detailni?section=shots&amp;order=s" class="btn btn-black">Kompletní tabulka</a>
+						 </div>
+					</div>
+				<div class="col-25">
+					<div class="banner banner-210" style="padding-top:40px"><a href="/hokejka-clanek/sledovani-nove-sezony-chance-ligy-opet-na-hokejka-tv/5050764"><img src="/files/images/84/210x300px-chl.jpg" alt="Hokejka TV" width="210" height="300"></a></div>
+				</div>
+		</div>
+	</div>
+</div>
+		</li>
+        <li>
+			<a href="/druha-liga?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" title="2. liga">2. liga</a>
+
+<div class="nav-main__sub">
+		<div class="nav-main__level2-container">
+			<div class="container">
+				<div class="row">
+					<div class="col-100">
+						<ul class="nav-main__level2">
+							<li><a href="/druha-liga">Zpravodajství</a></li>
+							<li>
+								<a href="/druha-liga/tymy">Týmy <i class="fa fa-chevron-down"></i></a>
+<div class="nav-main__sub">
+	<div>
+		<div class="col-100">
+			<div class="row nav-main__teams-container">
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/1920">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fbil.png&amp;w=76">
+						<span>Bílina</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/119">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fhcdc-logo.png&amp;w=76">
+						<span>Děčín</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/2654">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fdvur-kralove-logo.jpg&amp;w=76">
+						<span>Dvůr Králové</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/124">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fhavlickuv-brod-logo-16-17.png&amp;w=76">
+						<span>Havl. Brod</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/356">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2FDrtici-Hodonin-01-tiny.png&amp;w=76">
+						<span>Hodonín</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/2688">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Flogo-hronov-20.jpg&amp;w=76">
+						<span>Hronov</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/2685">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fcheb-logo-20.jpg&amp;w=76">
+						<span>Cheb</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/149">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fjablonec-logo-19.jpg&amp;w=76">
+						<span>Jablonec nad Nisou</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/150">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fklt.png&amp;w=76">
+						<span>Klatovy</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/88">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fkobra-praha-logo-2018.png&amp;w=76">
+						<span>Kobra Praha</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/99">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fkoprivnice.png&amp;w=76">
+						<span>Kopřivnice</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/2698">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fmost-logo-2018.png&amp;w=76">
+						<span>Most</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/2699">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fnova-paka-logo-19.jpg&amp;w=76">
+						<span>Nová Paka</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/96">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fnji.png&amp;w=76">
+						<span>Nový Jičín</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/4">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fopa.png&amp;w=76">
+						<span>Opava</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/40">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fbvmvy-pisek.png&amp;w=76">
+						<span>Písek</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/92">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fpribram.png&amp;w=76">
+						<span>Příbram</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/1919">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2F30.png&amp;w=76">
+						<span>Řisuty</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/121">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Ftab.png&amp;w=76">
+						<span>Tábor</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/1922">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Flogo-bak-trutnov.png&amp;w=76">
+						<span>Trutnov</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/1923">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fvlm.png&amp;w=76">
+						<span>Val. Meziříčí</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/2661">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Flogo-velke-mezirici.gif&amp;w=76">
+						<span>Velké Meziříčí</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/566">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Faqotec-orli-logo.png&amp;w=76">
+						<span>Znojmo</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/44">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Fzns.png&amp;w=76">
+						<span>Žďár n. Sázavou</span>
+					</a>
+				</div>
+			</div>
+		</div>
+	</div>
+</div>
+							</li>
+							<li><a href="/druha-liga/zapasy">Zápasy</a></li>
+							<li><a href="/druha-liga/table">Tabulky</a></li>
+							<li><a href="/druha-liga/statistika-tymu">Statistiky týmů</a></li>
+							<li><a href="/druha-liga/player-stats">Statistiky hráčů</a></li>
+						</ul>
+					</div>
+				</div>
+			</div>
+		</div>
+	<div class="container">
+		<div class="row">
+<div class="col-25">
+		<h2>Tabulka 2. liga – Sever</h2>
+		<table>
+			<tbody>
+				<tr>
+					<td>1.</td>
+					<td><a href="/klub/149?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Vlci Jablonec nad Nisou</a></td>
+					<td>9</td>
+				</tr>
+				<tr>
+					<td>2.</td>
+					<td><a href="/klub/2654?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Rodos Dvůr Králové</a></td>
+					<td>6</td>
+				</tr>
+				<tr>
+					<td>3.</td>
+					<td><a href="/klub/2688?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Wikov Hronov</a></td>
+					<td>6</td>
+				</tr>
+				<tr>
+					<td>4.</td>
+					<td><a href="/klub/119?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Děčín</a></td>
+					<td>6</td>
+				</tr>
+				<tr>
+					<td>5.</td>
+					<td><a href="/klub/2699?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">BK Nová Paka</a></td>
+					<td>3</td>
+				</tr>
+				<tr>
+					<td>6.</td>
+					<td><a href="/klub/1920?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Draci Bílina</a></td>
+					<td>3</td>
+				</tr>
+				<tr>
+					<td>7.</td>
+					<td><a href="/klub/1922?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Trutnov</a></td>
+					<td>3</td>
+				</tr>
+				<tr>
+					<td>8.</td>
+					<td><a href="/klub/2698?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">MOSTEČTÍ LVI</a></td>
+					<td>0</td>
+				</tr>
+			</tbody>
+		</table>
+		<div class="btn-container">
+			<a href="/druha-liga/table?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" class="btn btn-black">Podrobná tabulka</a>
+		</div>
+</div>					<div class="col-25">
+						<h2 class="m-b-10">Kanadské bodování</h2>
+<div class="row">
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/14821?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Daniel Svoboda</a>
+				<span>7 (3+4)</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/12993?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Jiří Půhoný</a>
+				<span>7 (3+4)</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/12160?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Jaroslav Moučka</a>
+				<span>7 (2+5)</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/10098119?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Alexander Yanchuk</a>
+				<span>6 (3+3)</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/11565?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Aleš Půlpán</a>
+				<span>5 (3+2)</span>
+			</p>
+		</div>
+	</div>
+</div>
+						<div class="btn-container">
+							<a href="/druha-liga/player-stats/detailni?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" class="btn btn-black">Kompletní statistiky</a>
+						</div>
+					</div>
+					<div class="col-25">
+						<h2 class="m-b-10">Tabulka střelců</h2>
+<div class="row">
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/14821?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Daniel Svoboda</a>
+				<span>3</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/12993?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Jiří Půhoný</a>
+				<span>3</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/10098119?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Alexander Yanchuk</a>
+				<span>3</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/11565?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Aleš Půlpán</a>
+				<span>3</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/9841?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Tomáš Bláha</a>
+				<span>3</span>
+			</p>
+		</div>
+	</div>
+</div>
+						 <div class="btn-container">
+							 <a href="/druha-liga/player-stats/detailni?section=shots&amp;order=s&amp;t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" class="btn btn-black">Kompletní tabulka</a>
+						 </div>
+					</div>
+		</div>
+	</div>
+</div>
+		</li>
+        <li>
+			<a href="/ullh?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" title="ULLH">ULLH</a>
+
+<div class="nav-main__sub">
+		<div class="nav-main__level2-container">
+			<div class="container">
+				<div class="row">
+					<div class="col-100">
+						<ul class="nav-main__level2">
+							<li><a href="/ullh">Zpravodajství</a></li>
+							<li>
+								<a href="/ullh/tymy">Týmy <i class="fa fa-chevron-down"></i></a>
+<div class="nav-main__sub">
+	<div>
+		<div class="col-100">
+			<div class="row nav-main__teams-container">
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/2716">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Flogo-akademici-plzen.png&amp;w=76">
+						<span>Akademici Plzeň</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/2717">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Flogo-black-dogs-budweis.png&amp;w=76">
+						<span>Black Dogs</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/2718">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Flogo-boostrava.png&amp;w=76">
+						<span>BO OSTRAVA</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/2719">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Flogo-vut-cavaliers.png&amp;w=76">
+						<span>Cavaliers</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/2720">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Flogo-prague-engineers.png&amp;w=76">
+						<span>Engineers</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/2721">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Flogo-hcmuni4.png&amp;w=76">
+						<span>HC MUNI</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/2726">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Flogo-north-wings-srpen2020.png&amp;w=76">
+						<span>HC NORTH WINGS</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/2722">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Flogo-hc-palacky-university.png&amp;w=76">
+						<span>HC Univerzita Palackého</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/2727">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Friders-pce-logo-zari2020.png&amp;w=76">
+						<span>Riders UP</span>
+					</a>
+				</div>
+				<div class="col-1_7 nav-main__team">
+					<a href="/klub/2723">
+						<img src="/min.php?file=%2Ffiles%2Flogos%2Flogo-uk-hockey2.png&amp;w=76">
+						<span>UK Hockey</span>
+					</a>
+				</div>
+			</div>
+		</div>
+	</div>
+</div>
+							</li>
+							<li><a href="/ullh/zapasy">Zápasy</a></li>
+							<li><a href="/ullh/table">Tabulky</a></li>
+													</ul>
+					</div>
+				</div>
+			</div>
+		</div>
+	<div class="container">
+		<div class="row">
+						
+			<div class="col-25">
+				
+				<div>
+		<h2>Tabulka Univerzitní liga – z. č.</h2>
+		<table>
+			<tbody>
+				<tr>
+					<td>1.</td>
+					<td><a href="/klub/2721">HC MUNI</a></td>
+					<td>6</td>
+				</tr>
+				<tr>
+					<td>2.</td>
+					<td><a href="/klub/2717">Black Dogs Budweis</a></td>
+					<td>6</td>
+				</tr>
+				<tr>
+					<td>3.</td>
+					<td><a href="/klub/2718">BO OSTRAVA Vítkovice Steel</a></td>
+					<td>5</td>
+				</tr>
+				<tr>
+					<td>4.</td>
+					<td><a href="/klub/2722">HC Univerzita Palackého v Olomouci</a></td>
+					<td>3</td>
+				</tr>
+				<tr>
+					<td>5.</td>
+					<td><a href="/klub/2720">Engineers Prague</a></td>
+					<td>3</td>
+				</tr>
+				<tr>
+					<td>6.</td>
+					<td><a href="/klub/2723">UK HOCKEY PRAGUE</a></td>
+					<td>1</td>
+				</tr>
+				<tr>
+					<td>7.</td>
+					<td><a href="/klub/2716">Akademici Plzeň</a></td>
+					<td>0</td>
+				</tr>
+				<tr>
+					<td>8.</td>
+					<td><a href="/klub/2719">VUT Cavaliers Brno</a></td>
+					<td>0</td>
+				</tr>
+				<tr>
+					<td>9.</td>
+					<td><a href="/klub/2726">HC North Wings</a></td>
+					<td>0</td>
+				</tr>
+				<tr>
+					<td>10.</td>
+					<td><a href="/klub/2727">Riders Univerzita Pardubice</a></td>
+					<td>0</td>
+				</tr>
+			</tbody>
+		</table>
+		<div class="btn-container">
+			<a href="/ullh/table" class="btn btn-black">Podrobná tabulka</a>
+		</div>
+</div>
+
+			</div>
+					<div class="col-25">
+						<h2 class="m-b-10">Kanadské bodování</h2>
+<div class="row">
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/14350">Tomáš Račický</a>
+				<span>5 (2+3)</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/3087343">Vojtěch Kadlic</a>
+				<span>5 (1+4)</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/10082149">Matěj Florek</a>
+				<span>4 (2+2)</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/50591">Lukáš Sajdl</a>
+				<span>4 (2+2)</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/36061">Radek Rolínek</a>
+				<span>4 (2+2)</span>
+			</p>
+		</div>
+	</div>
+</div>
+						<div class="btn-container">
+							<a href="/ullh/player-stats/detailni" class="btn btn-black">Kompletní statistiky</a>
+						</div>
+					</div>
+					<div class="col-25">
+						<h2 class="m-b-10">Tabulka střelců</h2>
+<div class="row">
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/20791">Martin Pěcha</a>
+				<span>3</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/16003">Marek Kovalski</a>
+				<span>3</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/28851">Tomáš Fabiánek</a>
+				<span>3</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/14350">Tomáš Račický</a>
+				<span>2</span>
+			</p>
+		</div>
+	</div>
+	<div class="person">
+				<div class="person__description">
+			<p>
+				<a href="/hrac/10082149">Matěj Florek</a>
+				<span>2</span>
+			</p>
+		</div>
+	</div>
+</div>
+						 <div class="btn-container">
+							 <a href="/ullh/player-stats/detailni?section=shots&amp;order=s" class="btn btn-black">Kompletní tabulka</a>
+						 </div>
+					</div>
+				<div class="col-25">
+					
+				</div>
+		</div>
+	</div>
+</div>
+		</li>
+        <li>
+			<a href="/mladez?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" title="Mládež">Mládež</a>
+<div class="nav-main__sub">
+		<div class="nav-main__level2-container">
+			<div class="container">
+				<div class="row">
+					<div class="col-100">
+						<ul class="nav-main__level2">
+<li><a href="/mladez/">Zpravodajství</a></li>
+<li><a href="/mladez/rozcestnik-souteze">Soutěže</a></li>
+<li><a href="/mladez/rozcestnik-reprezentace">Reprezentace</a></li> 
+<li><a href="/mladez/projekty-ch">Projekty ČH</a></li> 
+<li><a href="/mladez/clanky/category?category=334">Žákovský hokej</a></li> 
+<li><a href="/mladez/clanky/category?category=329">Trenéři a trénink</a></li> 
+<li><a href="/archiv-clanku/rubrika/327">Dívčí hokej</a></li>
+</ul>
+					</div>
+				</div>
+			</div>
+		</div>
+	<div class="container">
+		<div class="row">
+				<div class="col-25">
+<div class="box-league">
+	<div class="box-league__description">
+		<div class="box-league__logo">
+			<img alt="Reprezentace U20" src="/files/logos/novelogo-u20-1-.png">
+		</div>
+		<h3><a href="/mladez/liga/31">Reprezentace U20</a></h3>
+			<a href="/mladez/zapasy/31">Zápasy</a>
+			|
+			<a href="/mladez/table/31">Tabulky</a>
+			|
+		<a href="/mladez/player-stats/31">Statistiky</a>
+	</div>
+</div>
+<div class="box-league">
+	<div class="box-league__description">
+		<div class="box-league__logo">
+			<img alt="Reprezentace U19" src="/files/logos/novelogo-u19-1-.png">
+		</div>
+		<h3><a href="/mladez/liga/35">Reprezentace U19</a></h3>
+			<a href="/mladez/zapasy/35">Zápasy</a>
+			|
+			<a href="/mladez/table/35">Tabulky</a>
+			|
+		<a href="/mladez/player-stats/35">Statistiky</a>
+	</div>
+</div>
+<div class="box-league">
+	<div class="box-league__description">
+		<div class="box-league__logo">
+			<img alt="Reprezentace U18" src="/files/logos/novelogo-u18-1-.png">
+		</div>
+		<h3><a href="/mladez/liga/32">Reprezentace U18</a></h3>
+			<a href="/mladez/zapasy/32">Zápasy</a>
+			|
+			<a href="/mladez/table/32">Tabulky</a>
+			|
+		<a href="/mladez/player-stats/32">Statistiky</a>
+	</div>
+</div>
+<div class="box-league">
+	<div class="box-league__description">
+		<div class="box-league__logo">
+			<img alt="Reprezentace U17" src="/files/logos/novelogo-u17-1-.png">
+		</div>
+		<h3><a href="/mladez/liga/33">Reprezentace U17</a></h3>
+			<a href="/mladez/zapasy/33">Zápasy</a>
+			|
+			<a href="/mladez/table/33">Tabulky</a>
+			|
+		<a href="/mladez/player-stats/33">Statistiky</a>
+	</div>
+</div>
+<div class="box-league">
+	<div class="box-league__description">
+		<div class="box-league__logo">
+			<img alt="Reprezentace U16" src="/files/logos/novelogo-u16-1-.png">
+		</div>
+		<h3><a href="/mladez/liga/34">Reprezentace U16</a></h3>
+			<a href="/mladez/zapasy/34">Zápasy</a>
+			|
+			<a href="/mladez/table/34">Tabulky</a>
+			|
+		<a href="/mladez/player-stats/34">Statistiky</a>
+	</div>
+</div>
+									</div>
+				<div class="col-25">
+<div class="box-league">
+	<div class="box-league__description">
+		<h3><a href="/mladez/liga/55?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Juniorská liga akademií</a></h3>
+			<a href="/mladez/zapasy/55?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Zápasy</a>
+			|
+			<a href="/mladez/table/55?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Tabulky</a>
+			|
+		<a href="/mladez/player-stats/55?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Statistiky</a>
+	</div>
+</div>
+<div class="box-league">
+	<div class="box-league__description">
+		<h3><a href="/mladez/liga/4?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Extraliga juniorů</a></h3>
+			<a href="/mladez/zapasy/4?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Zápasy</a>
+			|
+			<a href="/mladez/table/4?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Tabulky</a>
+			|
+		<a href="/mladez/player-stats/4?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Statistiky</a>
+	</div>
+</div>
+<div class="box-league">
+	<div class="box-league__description">
+		<h3><a href="/mladez/liga/53?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Extraliga dorostu</a></h3>
+			<a href="/mladez/zapasy/53?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Zápasy</a>
+			|
+			<a href="/mladez/table/53?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Tabulky</a>
+			|
+		<a href="/mladez/player-stats/53?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Statistiky</a>
+	</div>
+</div>
+				</div>
+				<div class="col-25">
+					<div class="banner banner-210"><a href="https://www.pojdhrathokej.cz"><img src="https://www.hokej.cz/files/images/84/210-300-spravnysport-2020.jpg" alt="Pojď hrát hokej" style="margin-top:40px;" width="210" height="300"></a></div>
+				</div>
+		</div>
+	</div>
+</div>
+		</li>
+        <li>
+			<a href="/chl?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" title="CHL">CHL</a>
+	<div class="nav-main__sub">
+		<div class="nav-main__level2-container">
+			<div class="container">
+				<div class="row">
+					<div class="col-100">
+						<ul class="nav-main__level2">
+<li><a href="/chl/">Zpravodajství</a></li>
+<li><a href="/chl/tymy">Týmy</a></li>
+<li><a href="/chl/zapasy">Zápasy</a></li>
+<li><a href="/chl/table">Tabulky</a></li>
+</ul>
+					</div>
+				</div>
+			</div>
+		</div>
+		<div class="container">
+			<div class="row">
+				<div class="col-25">
+					<div style="padding-left:0px"><h2>České kluby v CHL</h2>
+<ul class="p-l-5">
+<li><a href="http://www.hc-ocelari.cz">HC Oceláři Třinec</a></li>
+<li><a href="http://www.hcskoda.cz">HC Škoda Plzeň</a></li>
+<li><a href="http://www.hcbilitygri.cz">Bílí Tygři Liberec</a></li>
+<li><a href="http://www.mountfieldhk.cz">Mountfield HK</a></li>
+</ul>
+<br>
+</div>
+				</div>
+				<div class="col-25">
+					<a href="http://www.championshockeyleague.net"><img src="/files/logos/chl-logo.png" alt="CHL logo" style="margin:25px 0 10px 0" width="200"></a>
+
+				</div>
+				<div class="col-25">
+					
+				</div>
+				<div class="col-25">
+					
+				</div>
+			</div>
+		</div>
+	</div>
+		</li>
+        <li>
+			<a href="/zensky-hokej?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" title="Ženy">Ženy</a>
+	<div class="nav-main__sub">
+		<div class="nav-main__level2-container">
+			<div class="container">
+				<div class="row">
+					<div class="col-100">
+						<ul class="nav-main__level2">
+<li><a href="/zensky-hokej">Zpravodajství</a></li>
+<li><a href="/zensky-hokej/liga/10">Soutěže žen</a></li>
+<li><a href="/zensky-hokej/stranka/5004228/section">Reprezentační akce</a></li>
+<li><a href="/zensky-hokej/stranka/5004236/section">Realizační týmy</a></li>
+<li><a href="/zensky-hokej/liga/39">Ženy "A"</a></li>
+<li><a href="/zensky-hokej/liga/40">Ženy U18</a></li>
+<li><a href="/zensky-hokej/liga/41">Ženy U16</a></li>
+<!-- <li><a href="/zensky-hokej/stranka/5004240/section">Žebříček IIHF</a></li> -->
+</ul>
+					</div>
+				</div>
+			</div>
+		</div>
+		<div class="container">
+			<div class="row">
+				<div class="col-25">
+					
+				</div>
+				<div class="col-25">
+					
+				</div>
+				<div class="col-25">
+					
+				</div>
+				<div class="col-25">
+					
+				</div>
+			</div>
+		</div>
+	</div>
+		</li>
+        <li>
+			<a href="/sledge-hokej?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" title="Para">Para</a>
+	<div class="nav-main__sub">
+		<div class="nav-main__level2-container">
+			<div class="container">
+				<div class="row">
+					<div class="col-100">
+						<ul class="nav-main__level2">
+<li><a href="/sledge-hokej">Zpravodajství</a></li>
+<li><a href="/sledge-hokej/liga/27">Para hokejová liga</a></li>
+<li><a href="/sledge-hokej/liga/42">Para reprezentace</a></li>
+<li><a href="/sledge-hokej/stranka/5004476/section">Reprezentační akce</a></li>
+<li><a href="/sledge-hokej/stranka/5004477/section">Realizační tým reprezentace</a></li>
+</ul>
+					</div>
+				</div>
+			</div>
+		</div>
+		<div class="container">
+			<div class="row">
+				<div class="col-25">
+					
+				</div>
+				<div class="col-25">
+					
+				</div>
+				<div class="col-25">
+					
+				</div>
+				<div class="col-25">
+					
+				</div>
+			</div>
+		</div>
+	</div>
+		</li>
+        <li>
+			<a href="/archiv-clanku/rubrika/114" title="Zahraničí">Zahraničí</a>
+		</li>
+        <li>
+			<a href="http://www.nhl.cz/" title="NHL.cz">NHL.cz</a>
+		</li>
+        <li>
+			<a href="/tv/hokejka/elh" title="TV">TV</a>
+		</li>
+        <li>
+			<a href="/hokejka/" title="Hokejka">Hokejka</a>
+		</li>
+        <li>
+			<a href="/historie?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" title="Historie">Historie</a>
+		</li>
+    </ul>
+</div>
+
+				<ul class="social">
+<li class="social-fb"><a href="https://www.facebook.com/hokejcz" target="_blank" title="facebook"><i class="fa fa-facebook"></i></a></li>
+<li class="social-tw"><a href="https://twitter.com/hokej_cz" target="_blank" title="twitter"><i class="fa fa-twitter"></i></a></li>
+<li class="social-yt"><a href="https://www.youtube.com/user/hokejcz" target="_blank" title="youtube"><i class="fa fa-youtube-play"></i></a></li>					<li class="social-rss"><a href="/rss/vse" target="_blank" title="rss"><i class="fa fa-rss"></i></a></li>
+				</ul>
+<div id="dl-menu" class="dl-menuwrapper">
+    <div class="dl-trigger">
+    	<div class="dl-bars">
+    		<span></span>
+    		<span></span>
+    		<span></span>
+    	</div>
+    	<span class="dl-title">menu</span>
+    </div>
+    <ul class="dl-menu">
+        <li>
+            <a href="/tipsport-extraliga?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" title="Tipsport ELH">Tipsport ELH</a>
+<ul class="dl-submenu"><li class="dl-back"><a href="#">zpět</a></li>
+	<li><a href="/tipsport-extraliga?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Zpravodajství</a></li>
+	<li>
+		<a href="/tipsport-extraliga/tymy?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Týmy</a>
+		<ul class="dl-submenu"><li class="dl-back"><a href="#">zpět</a></li>
+			<li><a href="/tipsport-extraliga/tymy?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Přehled</a></li>
+			<li><a href="/klub/1721?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Madeta Motor České Budějovice</a></li>
+			<li><a href="/klub/12?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Dynamo Pardubice</a></li>
+			<li><a href="/klub/1711?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Mountfield HK</a></li>
+			<li><a href="/klub/5?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Energie Karlovy Vary</a></li>
+			<li><a href="/klub/22?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Kometa Brno</a></li>
+			<li><a href="/klub/18?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Bílí Tygři Liberec</a></li>
+			<li><a href="/klub/823?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC VERVA Litvínov</a></li>
+			<li><a href="/klub/89?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">BK Mladá Boleslav</a></li>
+			<li><a href="/klub/11?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Oceláři Třinec</a></li>
+			<li><a href="/klub/24?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Olomouc</a></li>
+			<li><a href="/klub/917?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Škoda Plzeň</a></li>
+			<li><a href="/klub/8?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Sparta Praha</a></li>
+			<li><a href="/klub/7?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC VÍTKOVICE RIDERA</a></li>
+			<li><a href="/klub/562?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">PSG Berani Zlín</a></li>
+		</ul>
+	</li>
+	<li><a href="/tipsport-extraliga/zapasy?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Zápasy</a></li>
+	<li><a href="/tipsport-extraliga/table?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Tabulky</a></li>
+	<li><a href="/tipsport-extraliga/stats-center?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Centrum statistik</a></li>
+	<li><a href="/tipsport-extraliga/visualization?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Vizualizace</a></li>
+	<li><a href="/tipsport-extraliga/radegast-index?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Radegast Index</a></li>
+</ul>
+        </li>
+        <li>
+            <a href="/reprezentace?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" title="Reprezentace">Reprezentace</a>
+        </li>
+        <li>
+            <a href="/chance-liga?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" title="CHANCE LIGA">CHANCE LIGA</a>
+<ul class="dl-submenu"><li class="dl-back"><a href="#">zpět</a></li>
+	<li><a href="/chance-liga?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Zpravodajství</a></li>
+	<li>
+		<a href="/chance-liga/tymy?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Týmy</a>
+		<ul class="dl-submenu"><li class="dl-back"><a href="#">zpět</a></li>
+			<li><a href="/chance-liga/tymy?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Přehled</a></li>
+			<li><a href="/klub/91?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Benátky nad Jizerou</a></li>
+			<li><a href="/klub/132?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Frýdek-Místek</a></li>
+			<li><a href="/klub/714?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">AZ Heimstaden Havířov</a></li>
+			<li><a href="/klub/16?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Dukla Jihlava</a></li>
+			<li><a href="/klub/39?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">SK Trhači Kadaň</a></li>
+			<li><a href="/klub/828?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Rytíři Kladno</a></li>
+			<li><a href="/klub/117?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">SC Kolín</a></li>
+			<li><a href="/klub/500?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Stadion Litoměřice</a></li>
+			<li><a href="/klub/1927?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC RT TORAX Poruba 2011</a></li>
+			<li><a href="/klub/66?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">LHK Jestřábi Prostějov</a></li>
+			<li><a href="/klub/128?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC ZUBR Přerov</a></li>
+			<li><a href="/klub/9?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Slavia Praha</a></li>
+			<li><a href="/klub/123?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Baník Sokolov</a></li>
+			<li><a href="/klub/563?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">DRACI PARS ŠUMPERK</a></li>
+			<li><a href="/klub/17?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">SK Horácká Slavia Třebíč</a></li>
+			<li><a href="/klub/2611?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Hokej Ústí nad Labem s.r.o.</a></li>
+			<li><a href="/klub/1701?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Stadion Vrchlabí</a></li>
+			<li><a href="/klub/549?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">VHK ROBE Vsetín</a></li>
+		</ul>
+	</li>
+	<li><a href="/chance-liga/zapasy?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Zápasy</a></li>
+	<li><a href="/chance-liga/table?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Tabulky</a></li>
+	<li><a href="/chance-liga/stats-center?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Centrum statistik</a></li>
+</ul>        </li>
+        <li>
+            <a href="/druha-liga?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" title="2. liga">2. liga</a>
+<ul class="dl-submenu"><li class="dl-back"><a href="#">zpět</a></li>
+	<li><a href="/druha-liga?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Zpravodajství</a></li>
+	<li>
+		<a href="/druha-liga/tymy?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Týmy</a>
+		<ul class="dl-submenu"><li class="dl-back"><a href="#">zpět</a></li>
+			<li><a href="/druha-liga/tymy?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Přehled</a></li>
+			<li><a href="/klub/1920?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Draci Bílina</a></li>
+			<li><a href="/klub/119?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Děčín</a></li>
+			<li><a href="/klub/2654?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Rodos Dvůr Králové</a></li>
+			<li><a href="/klub/124?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">BK Havlíčkův Brod</a></li>
+			<li><a href="/klub/356?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">SHK Hodonín</a></li>
+			<li><a href="/klub/2688?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Wikov Hronov</a></li>
+			<li><a href="/klub/2685?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Stadion Cheb</a></li>
+			<li><a href="/klub/149?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Vlci Jablonec nad Nisou</a></li>
+			<li><a href="/klub/150?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">SHC Klatovy</a></li>
+			<li><a href="/klub/88?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Kobra Praha</a></li>
+			<li><a href="/klub/99?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC ISMM Kopřivnice</a></li>
+			<li><a href="/klub/2698?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">MOSTEČTÍ LVI</a></li>
+			<li><a href="/klub/2699?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">BK Nová Paka</a></li>
+			<li><a href="/klub/96?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Nový Jičín</a></li>
+			<li><a href="/klub/4?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Slezan Opava</a></li>
+			<li><a href="/klub/40?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">IHC Králové Písek</a></li>
+			<li><a href="/klub/92?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Příbram</a></li>
+			<li><a href="/klub/1919?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Řisuty</a></li>
+			<li><a href="/klub/121?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Tábor</a></li>
+			<li><a href="/klub/1922?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Trutnov</a></li>
+			<li><a href="/klub/1923?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Bobři Valašské Meziříčí</a></li>
+			<li><a href="/klub/2661?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HHK Velké Meziříčí</a></li>
+			<li><a href="/klub/566?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Aqotec Orli Znojmo</a></li>
+			<li><a href="/klub/44?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">SKLH Žďár nad Sázavou</a></li>
+		</ul>
+	</li>
+	<li><a href="/druha-liga/zapasy?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Zápasy</a></li>
+	<li><a href="/druha-liga/table?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Tabulky</a></li>
+	<li><a href="/druha-liga/statistika-tymu?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Statistiky týmů</a></li>
+	<li><a href="/druha-liga/player-stats?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Statistiky hráčů</a></li>
+</ul>        </li>
+        <li>
+            <a href="/ullh?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" title="ULLH">ULLH</a>
+<ul class="dl-submenu"><li class="dl-back"><a href="#">zpět</a></li>
+	<li><a href="/ullh?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Zpravodajství</a></li>
+	<li>
+		<a href="/ullh/tymy?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Týmy</a>
+		<ul class="dl-submenu"><li class="dl-back"><a href="#">zpět</a></li>
+			<li><a href="/ullh/tymy?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Přehled</a></li>
+			<li><a href="/klub/2716?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Akademici Plzeň</a></li>
+			<li><a href="/klub/2717?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Black Dogs Budweis</a></li>
+			<li><a href="/klub/2718?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">BO OSTRAVA Vítkovice Steel</a></li>
+			<li><a href="/klub/2719?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">VUT Cavaliers Brno</a></li>
+			<li><a href="/klub/2720?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Engineers Prague</a></li>
+			<li><a href="/klub/2721?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC MUNI</a></li>
+			<li><a href="/klub/2726?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC North Wings</a></li>
+			<li><a href="/klub/2722?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">HC Univerzita Palackého v Olomouci</a></li>
+			<li><a href="/klub/2727?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Riders Univerzita Pardubice</a></li>
+			<li><a href="/klub/2723?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">UK HOCKEY PRAGUE</a></li>
+		</ul>
+	</li>
+	<li><a href="/ullh/zapasy?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Zápasy</a></li>
+	<li><a href="/ullh/table?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Tabulky</a></li>
+	<li><a href="/ullh/stats-center?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2">Centrum statistik</a></li>
+</ul>        </li>
+        <li>
+            <a href="/mladez?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" title="Mládež">Mládež</a>
+        </li>
+        <li>
+            <a href="/chl?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" title="CHL">CHL</a>
+        </li>
+        <li>
+            <a href="/zensky-hokej?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" title="Ženy">Ženy</a>
+        </li>
+        <li>
+            <a href="/sledge-hokej?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" title="Para">Para</a>
+        </li>
+        <li>
+            <a href="/archiv-clanku/rubrika/114" title="Zahraničí">Zahraničí</a>
+        </li>
+        <li>
+            <a href="http://www.nhl.cz/" title="NHL.cz">NHL.cz</a>
+        </li>
+        <li>
+            <a href="/tv/hokejka/elh" title="TV">TV</a>
+        </li>
+        <li>
+            <a href="/hokejka/" title="Hokejka">Hokejka</a>
+        </li>
+        <li>
+            <a href="/historie?t=jzb8xxjdcacx3oaqbksb9o4scddrsag2lr1xkrvfmpaw7m23i6yfha2" title="Historie">Historie</a>
+        </li>
+    </ul>
+</div><!-- /dl-menuwrapper -->
+			</div>
+		</div>
+
+		<div class="bcg-sys-top">
+            <div class="banner"><!-- %New-on% -->
+<!-- %Default%;%ID%:1x1;%Banner%:Leaderboard-1 -->
+
+<div class="ads"><div class="ads-in" id="Leaderboard-1"><script> if(typeof _SAS === 'object') _SAS.ad({site:'Hokej', area:'Ostatni', pos:'Leaderboard-1', type:'category'}); </script></div></div>
+
+
+
+</div>
+		</div>
+
+<div id="snippet--login"></div>
+<div id="snippet--register"></div>
+
+<div id="snippet--fbRegister"></div>
+
+<div id="snippet--editForm"></div>
+<div id="snippet--passwordRecoveryForm"></div>
+<div id="snippet--passwordUpdateForm"></div>
+<div id="snippet--voteSuccess"></div>
+<div id="snippet--thanksForVote"></div>
+<div id="snippet--registerSuccess"></div>
+<div id="snippet--accountActivateSuccess"></div>
+<div id="snippet--passwordRecoverySuccess"></div>
+<div id="snippet--editSuccess"></div>
+<div id="snippet--unsubscribeEmailSuccess"></div>
+
+				<div class="content-wrapper">
+			<div class="container">
+
+
+				<div class="header-bottom ">
+					<div class="search-all">
+						<form action="/vyhledavani" method="get" novalidate="">
+							<input type="text" placeholder="Vyhledat na webu …" name="q">
+							<input type="submit" value="">
+						</form>
+					</div>
+<div class="search-person">
+	<form action="/hrac/23461/career?stats-section=all" method="post" id="frm-playerSearch-form" class="player-search" autocomplete="off" novalidate="">
+		
+		<input type="text" name="q" id="frm-playerSearch-form-q" placeholder="Vyhledat hráče...">
+		<label for="frm-playerSearch-form-q" class="label-like-submit">Jméno</label>
+		<input type="submit" name="search" value="Hledat" class="input-hide">
+	<input type="hidden" name="do" value="playerSearch-form-submit"><!--[if IE]><input type=IEbug disabled style="display:none"><![endif]-->
+</form>
+
+<div id="snippet-playerSearch-"></div></div>
+					<div class="user-section">
+
+												
+							<a href="http://www.hokej.cz/hokejka-ticket" class="user-section-logo" target="_blank">
+								<img src="/images/hokejkaTicket_button.svg" title="Komunitní projekt českého hokeje" alt="Komunitní projekt českého hokeje">
+							</a>
+
+			            <ul class="list-unstyled">
+							<li class="user-section-sign"><a href="/hokejka?login=true">Přihlášení <i class="fa fa-sign-in"></i></a></li>
+							<li class="user-section-registration"><a href="/hokejka/kabina/register-new-member">Registrace <i class="fa fa-pencil"></i></a></li>
+			            </ul>
+					</div>
+
+									</div>
+
+    <div class="container">
+<div class="row m-t-30 person-info text-center">
+    <div class="col-100 person-info__image">
+        <img src="/static/images/hrac/kr/kral.jpg" alt="Fotka Filip Král" class="circle" width="140" height="140">
+    </div>
+    <h1>Filip Král</h1>
+    <ul class="m-t-0 m-b-20 p-l-0">
+        <li class="person-info-club d-inline-block">
+            <a class="d-block" href="/klub/128">HC ZUBR Přerov</a></li>
+    </ul>
+    <div class="col-100 m-b-20">
+        <div class="row text-center">
+            <div class="person-info__data d-inline-block">
+                <h2 class="h4 person-info-title text-uppercase">narozen</h2>
+                <span>20.10.1999</span>
+            </div>
+                <div class="person-info__data d-inline-block">
+                    <h2 class="h4 person-info-title text-uppercase">věk</h2>
+                    <span>21 let</span>
+                </div>
+            <div class="person-info__data d-inline-block">
+                <h2 class="h4 person-info-title text-uppercase">výška</h2>
+                <span>182 cm</span>
+            </div>
+            <div class="person-info__data d-inline-block">
+                <h2 class="h4 person-info-title text-uppercase">váha</h2>
+                <span>80 kg</span>
+            </div>
+            <div class="person-info__data d-inline-block">
+                <h2 class="h4 person-info-title text-uppercase">post</h2>
+                <span class="before-last">obránce</span>
+            </div>
+            <div class="person-info__data d-inline-block">
+                <h2 class="h4 person-info-title text-uppercase">hůl</h2>
+                <span class="last">levá</span>
+            </div>
+        </div>
+    </div>
+
+        <div class="person-info-stats d-inline-block p-20">
+            <h2 class="c-white b-b-0 m-t-0 m-b-20">Tipsport extraliga 2020-2021</h2>
+
+            <div class="person-info-stats-box d-inline-block m-b-10">
+                <div class="person-info-stats-title font-weight-bold c-white text-uppercase">Zápasy</div>
+                <div class="person-info-stats-value font-weight-bold c-white m-t-5">3</div>
+            </div>
+
+                <div class="person-info-stats-box d-inline-block m-b-10">
+                    <div class="person-info-stats-title font-weight-bold c-white text-uppercase">Góly</div>
+                    <div class="person-info-stats-value font-weight-bold c-white m-t-5">0</div>
+                </div>
+                <div class="person-info-stats-box d-inline-block m-b-10">
+                    <div class="person-info-stats-title font-weight-bold c-white text-uppercase">Asistence</div>
+                    <div class="person-info-stats-value font-weight-bold c-white m-t-5">2</div>
+                </div>
+                <div class="person-info-stats-box d-inline-block m-b-10">
+                    <div class="person-info-stats-title font-weight-bold c-white text-uppercase">Body</div>
+                    <div class="person-info-stats-value font-weight-bold c-white m-t-5">2</div>
+                </div>
+                <div class="person-info-stats-box d-inline-block m-b-10">
+                    <div class="person-info-stats-title font-weight-bold c-white text-uppercase">+/-</div>
+                    <div class="person-info-stats-value font-weight-bold c-white m-t-5">2</div>
+                </div>
+                <div class="person-info-stats-box d-inline-block m-b-10">
+                    <div class="person-info-stats-title font-weight-bold c-white text-uppercase">TM</div>
+                    <div class="person-info-stats-value font-weight-bold c-white m-t-5">0</div>
+                </div>
+                <div class="person-info-stats-box d-inline-block m-b-10">
+                    <div class="person-info-stats-title font-weight-bold c-white text-uppercase">R.Index</div>
+                    <div class="person-info-stats-value font-weight-bold c-white m-t-5">6</div>
+                </div>
+        </div>
+</div>
+
+<div class="nav-tab">
+    <div class="container row">
+        <div class="col-100">
+            <ul class="nav-tab__list">
+				<li>
+					<a href="/hrac/23461" class="filter-link">Přehled</a>
+				</li>
+				<li>
+					<a href="/hrac/23461/career" class="active filter-link">Kariéra</a>
+				</li>
+				<li>
+					<a href="/hrac/23461/match" class="filter-link">Zápasy</a>
+				</li>
+			</ul>
+		</div>
+	</div>
+</div>    </div>
+
+
+
+
+
+
+
+
+
+
+<div class="container p-b-20"> 
+	<div>
+		<ul class="nav-filter">
+			<li><a href="/hrac/23461/career?stats-section=all&amp;t=224&amp;do=stats-section" class="active">Vše</a></li>
+			<li><a href="/hrac/23461/career?stats-section=junior&amp;t=224&amp;do=stats-section">Juniorské soutěže</a></li>
+			<li><a href="/hrac/23461/career?stats-section=senior&amp;t=224&amp;do=stats-section">Seniorské soutěže</a></li>
+		</ul>
+	</div>
+<div class="row m-b-5">
+	<div class="col-60">
+		<h2 class="b-b-0">Klubová kariéra</h2>
+	</div>
+	<div class="col-40">
+		<button href="" class="btn btn-small btn-white pull-right" onclick="tgl('#table--1 .table-stats__highlight')">Zobrazit části soutěží</button>
+	</div>
+</div>
+    <div class="row">
+<div class="table-stats-wrapper">
+	<table class="table-stats" id="table--1" style="min-widht: 900px;">
+
+<thead>
+	<tr>
+		<th></th>
+		<th>soutěž</th>
+		<th>klub</th>
+		<th class="hasqtip" data-content="Zápasy" data-hasqtip="0">z</th>
+		<th class="hasqtip" data-content="Góly hráče" data-hasqtip="1">g</th>
+
+		<th class="hasqtip" data-content="Asistence" data-hasqtip="2">a</th>
+		<th class="hasqtip" data-content="Body (G + A)" data-hasqtip="3">b</th>
+		<th class="hasqtip" data-content="Hodnocení hráče za pozitivní a negativní účast na ledě" data-hasqtip="4">+/-</th>
+		<th class="hasqtip" data-content="Hodnocení hráče za pozitivní účast na ledě" data-hasqtip="5">+</th>
+
+		<th class="hasqtip" data-content="Hodnocení hráče za negativní účast na ledě" data-hasqtip="6">-</th>
+		<th class="hasqtip" data-content="Trestné minuty" data-hasqtip="7">tm</th>
+		<th class="hasqtip" data-content="Góly vítězné" data-hasqtip="8">gv</th>
+		<th class="hasqtip" data-content="Góly v přesilovkách" data-hasqtip="9">gp</th>
+
+		<th class="hasqtip" data-content="Góly v oslabení" data-hasqtip="10">go</th>
+		<th></th>
+	</tr>
+</thead>
+		<tbody>
+                    
+					<tr>
+						<th class="table-stats__first" rowspan="2">2012-2013</th>
+
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/6?competition=1140&amp;t=224">
+            ELIOD ELMD
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/6?competition=1140&amp;teamId=604&amp;t=224">MD Kometa Brno</a></td>
+<td class="tbl-c-3">1</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">0</td>
+<td class="tbl-c-6">0</td>
+<td class="tbl-c-7">-1</td>
+<td class="tbl-c-8">0</td>
+<td class="tbl-c-9">1</td>
+
+<td class="tbl-c-10">0</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/6?competition=1140&amp;t=224">
+            ELIOD ELMD – S
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/6?competition=1140&amp;teamId=604&amp;t=224">MD Kometa Brno</a></td>
+<td class="tbl-c-3">1</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">0</td>
+<td class="tbl-c-6">0</td>
+<td class="tbl-c-7">-1</td>
+<td class="tbl-c-8">0</td>
+<td class="tbl-c-9">1</td>
+
+<td class="tbl-c-10">0</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+		</tbody>
+		<tbody>
+                    
+					<tr>
+						<th class="table-stats__first" rowspan="3">2013-2014</th>
+
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/6?competition=2611&amp;t=224">
+            ELIOD ELMD
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/6?competition=2611&amp;teamId=604&amp;t=224">MD Kometa Brno</a></td>
+<td class="tbl-c-3">33</td>
+<td class="tbl-c-4">6</td>
+<td class="tbl-c-5">5</td>
+<td class="tbl-c-6">11</td>
+<td class="tbl-c-7">9</td>
+<td class="tbl-c-8">21</td>
+<td class="tbl-c-9">12</td>
+
+<td class="tbl-c-10">6</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/6?competition=2611&amp;t=224">
+            ELIOD ELMD – S
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/6?competition=2611&amp;teamId=604&amp;t=224">MD Kometa Brno</a></td>
+<td class="tbl-c-3">33</td>
+<td class="tbl-c-4">6</td>
+<td class="tbl-c-5">5</td>
+<td class="tbl-c-6">11</td>
+<td class="tbl-c-7">9</td>
+<td class="tbl-c-8">21</td>
+<td class="tbl-c-9">12</td>
+
+<td class="tbl-c-10">6</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+                    
+					<tr>
+
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/6?competition=3821&amp;t=224">
+            ELIOD ELMD – play off
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/6?competition=3821&amp;teamId=604&amp;t=224">MD Kometa Brno</a></td>
+<td class="tbl-c-3">4</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">3</td>
+<td class="tbl-c-6">3</td>
+<td class="tbl-c-7">2</td>
+<td class="tbl-c-8">3</td>
+<td class="tbl-c-9">1</td>
+
+<td class="tbl-c-10">0</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+		</tbody>
+		<tbody>
+                    
+					<tr>
+						<th class="table-stats__first" rowspan="8">2014-2015</th>
+
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/5?competition=5717&amp;t=224">
+            RAKO ELSD
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/5?competition=5717&amp;teamId=589&amp;t=224">SD Kometa Brno</a></td>
+<td class="tbl-c-3">8</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">4</td>
+<td class="tbl-c-6">4</td>
+<td class="tbl-c-7">10</td>
+<td class="tbl-c-8">13</td>
+<td class="tbl-c-9">3</td>
+
+<td class="tbl-c-10">0</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/5?competition=5717&amp;t=224">
+            RAKO ELSD – V
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/5?competition=5717&amp;teamId=589&amp;t=224">SD Kometa Brno</a></td>
+<td class="tbl-c-3">5</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">2</td>
+<td class="tbl-c-6">2</td>
+<td class="tbl-c-7">8</td>
+<td class="tbl-c-8">10</td>
+<td class="tbl-c-9">2</td>
+
+<td class="tbl-c-10">0</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/5?competition=5717&amp;t=224">
+            RAKO ELSD II – A
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/5?competition=5717&amp;teamId=589&amp;t=224">SD Kometa Brno</a></td>
+<td class="tbl-c-3">3</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">2</td>
+<td class="tbl-c-6">2</td>
+<td class="tbl-c-7">2</td>
+<td class="tbl-c-8">3</td>
+<td class="tbl-c-9">1</td>
+
+<td class="tbl-c-10">0</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+                    
+					<tr>
+
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/6?competition=5552&amp;t=224">
+            ELIOD ELMD – play off – B
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/6?competition=5552&amp;teamId=604&amp;t=224">MD Kometa Brno</a></td>
+<td class="tbl-c-3">4</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">4</td>
+<td class="tbl-c-6">4</td>
+<td class="tbl-c-7">7</td>
+<td class="tbl-c-8">7</td>
+<td class="tbl-c-9">0</td>
+
+<td class="tbl-c-10">8</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+                    
+					<tr>
+
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/6?competition=5555&amp;t=224">
+            ELIOD ELMD – závěrečný turnaj
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/6?competition=5555&amp;teamId=604&amp;t=224">MD Kometa Brno</a></td>
+<td class="tbl-c-3">3</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">0</td>
+<td class="tbl-c-6">0</td>
+<td class="tbl-c-7">1</td>
+<td class="tbl-c-8">2</td>
+<td class="tbl-c-9">1</td>
+
+<td class="tbl-c-10">0</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+                    
+					<tr>
+
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/6?competition=5716&amp;t=224">
+            ELIOD ELMD
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/6?competition=5716&amp;teamId=604&amp;t=224">MD Kometa Brno</a></td>
+<td class="tbl-c-3">31</td>
+<td class="tbl-c-4">12</td>
+<td class="tbl-c-5">31</td>
+<td class="tbl-c-6">43</td>
+<td class="tbl-c-7">37</td>
+<td class="tbl-c-8">50</td>
+<td class="tbl-c-9">13</td>
+
+<td class="tbl-c-10">12</td>
+<td class="tbl-c-11">3</td>
+<td class="tbl-c-12">4</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/6?competition=5716&amp;t=224">
+            ELIOD ELMD – S
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/6?competition=5716&amp;teamId=604&amp;t=224">MD Kometa Brno</a></td>
+<td class="tbl-c-3">16</td>
+<td class="tbl-c-4">2</td>
+<td class="tbl-c-5">20</td>
+<td class="tbl-c-6">22</td>
+<td class="tbl-c-7">16</td>
+<td class="tbl-c-8">25</td>
+<td class="tbl-c-9">9</td>
+
+<td class="tbl-c-10">12</td>
+<td class="tbl-c-11">1</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/6?competition=5716&amp;t=224">
+            ELIOD ELMD – o postup – sk. B
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/6?competition=5716&amp;teamId=604&amp;t=224">MD Kometa Brno</a></td>
+<td class="tbl-c-3">15</td>
+<td class="tbl-c-4">10</td>
+<td class="tbl-c-5">11</td>
+<td class="tbl-c-6">21</td>
+<td class="tbl-c-7">21</td>
+<td class="tbl-c-8">25</td>
+<td class="tbl-c-9">4</td>
+
+<td class="tbl-c-10">0</td>
+<td class="tbl-c-11">2</td>
+<td class="tbl-c-12">4</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+		</tbody>
+		<tbody>
+                    
+					<tr>
+						<th class="table-stats__first" rowspan="4">2015-2016</th>
+
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/5?competition=5718&amp;t=224">
+            NOEN ELSD
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/5?competition=5718&amp;teamId=589&amp;t=224">SD Kometa Brno</a></td>
+<td class="tbl-c-3">40</td>
+<td class="tbl-c-4">11</td>
+<td class="tbl-c-5">29</td>
+<td class="tbl-c-6">40</td>
+<td class="tbl-c-7">25</td>
+<td class="tbl-c-8">57</td>
+<td class="tbl-c-9">32</td>
+
+<td class="tbl-c-10">57</td>
+<td class="tbl-c-11">2</td>
+<td class="tbl-c-12">5</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/5?competition=5718&amp;t=224">
+            NOEN ELSD – S
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/5?competition=5718&amp;teamId=589&amp;t=224">SD Kometa Brno</a></td>
+<td class="tbl-c-3">27</td>
+<td class="tbl-c-4">8</td>
+<td class="tbl-c-5">21</td>
+<td class="tbl-c-6">29</td>
+<td class="tbl-c-7">27</td>
+<td class="tbl-c-8">43</td>
+<td class="tbl-c-9">16</td>
+
+<td class="tbl-c-10">18</td>
+<td class="tbl-c-11">2</td>
+<td class="tbl-c-12">4</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/5?competition=5718&amp;t=224">
+            NOEN ELSD II – A
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/5?competition=5718&amp;teamId=589&amp;t=224">SD Kometa Brno</a></td>
+<td class="tbl-c-3">13</td>
+<td class="tbl-c-4">3</td>
+<td class="tbl-c-5">8</td>
+<td class="tbl-c-6">11</td>
+<td class="tbl-c-7">-2</td>
+<td class="tbl-c-8">14</td>
+<td class="tbl-c-9">16</td>
+
+<td class="tbl-c-10">39</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">1</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+                    
+					<tr>
+
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/5?competition=5782&amp;t=224">
+            NOEN ELSD – play off
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/5?competition=5782&amp;teamId=589&amp;t=224">SD Kometa Brno</a></td>
+<td class="tbl-c-3">10</td>
+<td class="tbl-c-4">1</td>
+<td class="tbl-c-5">3</td>
+<td class="tbl-c-6">4</td>
+<td class="tbl-c-7">8</td>
+<td class="tbl-c-8">13</td>
+<td class="tbl-c-9">5</td>
+
+<td class="tbl-c-10">4</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+		</tbody>
+		<tbody>
+                    
+					<tr>
+						<th class="table-stats__first" rowspan="5">2016-2017</th>
+
+<td class="tbl-c-1">
+		<a href="/tipsport-extraliga/stats-center?competition=5821&amp;season=2016&amp;t=224">
+			Tipsport extraliga
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/tipsport-extraliga/stats-center?competition=5821&amp;team=22&amp;season=2016&amp;state=&amp;t=224">HC Kometa Brno</a></td>
+<td class="tbl-c-3">23</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">2</td>
+<td class="tbl-c-6">2</td>
+<td class="tbl-c-7">3</td>
+<td class="tbl-c-8">6</td>
+<td class="tbl-c-9">3</td>
+
+<td class="tbl-c-10">2</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+                    
+					<tr>
+
+<td class="tbl-c-1">
+		<a href="/tipsport-extraliga/stats-center?competition=5967&amp;season=2016&amp;t=224">
+			Generali play off
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/tipsport-extraliga/stats-center?competition=5967&amp;team=22&amp;season=2016&amp;state=&amp;t=224">HC Kometa Brno</a></td>
+<td class="tbl-c-3">5</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">0</td>
+<td class="tbl-c-6">0</td>
+<td class="tbl-c-7">0</td>
+<td class="tbl-c-8">0</td>
+<td class="tbl-c-9">0</td>
+
+<td class="tbl-c-10">0</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+                    
+					<tr>
+
+<td class="tbl-c-1">
+		<a href="/chance-liga/stats-center?competition=5822&amp;season=2016&amp;t=224">
+			WSM Liga
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/chance-liga/stats-center?competition=5822&amp;team=17&amp;season=2016&amp;state=&amp;t=224">SK Horácká Slavia Třebíč</a></td>
+<td class="tbl-c-3">12</td>
+<td class="tbl-c-4">1</td>
+<td class="tbl-c-5">1</td>
+<td class="tbl-c-6">2</td>
+<td class="tbl-c-7">-1</td>
+<td class="tbl-c-8">6</td>
+<td class="tbl-c-9">7</td>
+
+<td class="tbl-c-10">6</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+                    
+					<tr>
+
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/4?competition=5864&amp;t=224">
+            DHL ELJ
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/4?competition=5864&amp;teamId=261&amp;t=224">J Kometa Brno</a></td>
+<td class="tbl-c-3">13</td>
+<td class="tbl-c-4">4</td>
+<td class="tbl-c-5">6</td>
+<td class="tbl-c-6">10</td>
+<td class="tbl-c-7">9</td>
+<td class="tbl-c-8">17</td>
+<td class="tbl-c-9">8</td>
+
+<td class="tbl-c-10">10</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">1</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+                    
+					<tr>
+
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/4?competition=6005&amp;t=224">
+            DHL ELJ – play off
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/4?competition=6005&amp;teamId=261&amp;t=224">J Kometa Brno</a></td>
+<td class="tbl-c-3">6</td>
+<td class="tbl-c-4">3</td>
+<td class="tbl-c-5">7</td>
+<td class="tbl-c-6">10</td>
+<td class="tbl-c-7">5</td>
+<td class="tbl-c-8">7</td>
+<td class="tbl-c-9">2</td>
+
+<td class="tbl-c-10">2</td>
+<td class="tbl-c-11">1</td>
+<td class="tbl-c-12">2</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+		</tbody>
+		<tbody>
+                    
+					<tr>
+						<th class="table-stats__first" rowspan="2">2017-2018</th>
+
+<td class="tbl-c-1">
+		<a href="/tipsport-extraliga/stats-center?competition=6026&amp;season=2017&amp;t=224">
+			Tipsport extraliga
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/tipsport-extraliga/stats-center?competition=6026&amp;team=22&amp;season=2017&amp;state=&amp;t=224">HC Kometa Brno</a></td>
+<td class="tbl-c-3">4</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">0</td>
+<td class="tbl-c-6">0</td>
+<td class="tbl-c-7">1</td>
+<td class="tbl-c-8">1</td>
+<td class="tbl-c-9">0</td>
+
+<td class="tbl-c-10">0</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+                    
+					<tr>
+
+<td class="tbl-c-1">
+		<a href="/chance-liga/stats-center?competition=6045&amp;season=2017&amp;t=224">
+			WSM Liga
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/chance-liga/stats-center?competition=6045&amp;team=17&amp;season=2017&amp;state=&amp;t=224">SK Horácká Slavia Třebíč</a></td>
+<td class="tbl-c-3">3</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">0</td>
+<td class="tbl-c-6">0</td>
+<td class="tbl-c-7">2</td>
+<td class="tbl-c-8">4</td>
+<td class="tbl-c-9">2</td>
+
+<td class="tbl-c-10">2</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+		</tbody>
+		<tbody>
+                    
+					<tr>
+						<th class="table-stats__first" rowspan="4">2020-2021</th>
+
+<td class="tbl-c-1">
+		<a href="/tipsport-extraliga/stats-center?competition=6630&amp;season=2020&amp;t=224">
+			Tipsport extraliga
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/tipsport-extraliga/stats-center?competition=6630&amp;team=22&amp;season=2020&amp;state=&amp;t=224">HC Kometa Brno</a></td>
+<td class="tbl-c-3">3</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">2</td>
+<td class="tbl-c-6">2</td>
+<td class="tbl-c-7">2</td>
+<td class="tbl-c-8">3</td>
+<td class="tbl-c-9">1</td>
+
+<td class="tbl-c-10">0</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+                    
+					<tr>
+
+<td class="tbl-c-1">
+		<a href="/tipsport-extraliga/stats-center?competition=6633&amp;season=2020&amp;t=224">
+			Generali Česká Cup
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/tipsport-extraliga/stats-center?competition=6633&amp;team=128&amp;season=2020&amp;state=&amp;t=224">HC ZUBR Přerov</a></td>
+<td class="tbl-c-3">3</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">0</td>
+<td class="tbl-c-6">0</td>
+<td class="tbl-c-7">0</td>
+<td class="tbl-c-8">1</td>
+<td class="tbl-c-9">1</td>
+
+<td class="tbl-c-10">2</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/tipsport-extraliga/stats-center?competition=6633&amp;season=2020&amp;t=224">
+			Generali Česká Cup – sk. D
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/tipsport-extraliga/stats-center?competition=6633&amp;team=128&amp;season=2020&amp;state=&amp;t=224">HC ZUBR Přerov</a></td>
+<td class="tbl-c-3">3</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">0</td>
+<td class="tbl-c-6">0</td>
+<td class="tbl-c-7">0</td>
+<td class="tbl-c-8">1</td>
+<td class="tbl-c-9">1</td>
+
+<td class="tbl-c-10">2</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+                    
+					<tr>
+
+<td class="tbl-c-1">
+		<a href="/chance-liga/stats-center?competition=6631&amp;season=2020&amp;t=224">
+			CHANCE LIGA
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/chance-liga/stats-center?competition=6631&amp;team=128&amp;season=2020&amp;state=&amp;t=224">HC ZUBR Přerov</a></td>
+<td class="tbl-c-3">7</td>
+<td class="tbl-c-4">2</td>
+<td class="tbl-c-5">8</td>
+<td class="tbl-c-6">10</td>
+<td class="tbl-c-7">4</td>
+<td class="tbl-c-8">9</td>
+<td class="tbl-c-9">5</td>
+
+<td class="tbl-c-10">4</td>
+<td class="tbl-c-11">1</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">1</td>
+<td></td>
+					</tr>
+		</tbody>
+<tbody>
+	<tr>
+		
+</tr></tbody><thead>
+	<tr>
+		<th></th>
+		<th>soutěž</th>
+		<th>klub</th>
+		<th class="hasqtip" data-content="Zápasy" data-hasqtip="11">z</th>
+		<th class="hasqtip" data-content="Góly hráče" data-hasqtip="12">g</th>
+
+		<th class="hasqtip" data-content="Asistence" data-hasqtip="13">a</th>
+		<th class="hasqtip" data-content="Body (G + A)" data-hasqtip="14">b</th>
+		<th class="hasqtip" data-content="Hodnocení hráče za pozitivní a negativní účast na ledě" data-hasqtip="15">+/-</th>
+		<th class="hasqtip" data-content="Hodnocení hráče za pozitivní účast na ledě" data-hasqtip="16">+</th>
+
+		<th class="hasqtip" data-content="Hodnocení hráče za negativní účast na ledě" data-hasqtip="17">-</th>
+		<th class="hasqtip" data-content="Trestné minuty" data-hasqtip="18">tm</th>
+		<th class="hasqtip" data-content="Góly vítězné" data-hasqtip="19">gv</th>
+		<th class="hasqtip" data-content="Góly v přesilovkách" data-hasqtip="20">gp</th>
+
+		<th class="hasqtip" data-content="Góly v oslabení" data-hasqtip="21">go</th>
+		<th></th>
+	</tr>
+</thead>
+
+		<tbody><tr><th class="table-stats__first" rowspan="6">Vše</th>
+<td class="tbl-c-1">
+		<a href="/tipsport-extraliga/stats-center?t=224">
+			Generali Česká Cup
+		</a>
+</td>
+		<td class="tbl-c-2">1 klub</td>
+
+<td class="tbl-c-3">3</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">0</td>
+<td class="tbl-c-6">0</td>
+<td class="tbl-c-7">0</td>
+<td class="tbl-c-8">1</td>
+<td class="tbl-c-9">1</td>
+
+<td class="tbl-c-10">2</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+	</tr>
+	<tr>
+		
+<td class="tbl-c-1">
+		<a href="/tipsport-extraliga/stats-center?t=224">
+			Extraliga
+		</a>
+</td>
+		<td class="tbl-c-2">1 klub</td>
+
+<td class="tbl-c-3">35</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">4</td>
+<td class="tbl-c-6">4</td>
+<td class="tbl-c-7">6</td>
+<td class="tbl-c-8">10</td>
+<td class="tbl-c-9">4</td>
+
+<td class="tbl-c-10">2</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+	</tr>
+	<tr>
+		
+<td class="tbl-c-1">
+		<a href="/chance-liga/stats-center?t=224">
+			CHANCE LIGA
+		</a>
+</td>
+		<td class="tbl-c-2">2 kluby</td>
+
+<td class="tbl-c-3">22</td>
+<td class="tbl-c-4">3</td>
+<td class="tbl-c-5">9</td>
+<td class="tbl-c-6">12</td>
+<td class="tbl-c-7">5</td>
+<td class="tbl-c-8">19</td>
+<td class="tbl-c-9">14</td>
+
+<td class="tbl-c-10">12</td>
+<td class="tbl-c-11">1</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">1</td>
+<td></td>
+	</tr>
+	<tr>
+		
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/4?t=224">
+            Extraliga juniorů
+		</a>
+</td>
+		<td class="tbl-c-2">1 klub</td>
+
+<td class="tbl-c-3">19</td>
+<td class="tbl-c-4">7</td>
+<td class="tbl-c-5">13</td>
+<td class="tbl-c-6">20</td>
+<td class="tbl-c-7">14</td>
+<td class="tbl-c-8">24</td>
+<td class="tbl-c-9">10</td>
+
+<td class="tbl-c-10">12</td>
+<td class="tbl-c-11">1</td>
+<td class="tbl-c-12">3</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+	</tr>
+	<tr>
+		
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/5?t=224">
+            Extraliga staršího dorostu
+		</a>
+</td>
+		<td class="tbl-c-2">1 klub</td>
+
+<td class="tbl-c-3">58</td>
+<td class="tbl-c-4">12</td>
+<td class="tbl-c-5">36</td>
+<td class="tbl-c-6">48</td>
+<td class="tbl-c-7">43</td>
+<td class="tbl-c-8">83</td>
+<td class="tbl-c-9">40</td>
+
+<td class="tbl-c-10">61</td>
+<td class="tbl-c-11">2</td>
+<td class="tbl-c-12">5</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+	</tr>
+	<tr>
+		
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/6?t=224">
+            Extraliga mladšího dorostu
+		</a>
+</td>
+		<td class="tbl-c-2">1 klub</td>
+
+<td class="tbl-c-3">76</td>
+<td class="tbl-c-4">18</td>
+<td class="tbl-c-5">43</td>
+<td class="tbl-c-6">61</td>
+<td class="tbl-c-7">55</td>
+<td class="tbl-c-8">83</td>
+<td class="tbl-c-9">28</td>
+
+<td class="tbl-c-10">26</td>
+<td class="tbl-c-11">3</td>
+<td class="tbl-c-12">4</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+	</tr>
+</tbody>
+	</table>
+</div>
+    </div>
+<div class="row m-b-5">
+	<div class="col-60">
+		<h2 class="b-b-0">Reprezentační kariéra</h2>
+	</div>
+	<div class="col-40">
+		<button href="" class="btn btn-small btn-white pull-right" onclick="tgl('#table--2 .table-stats__highlight')">Zobrazit části soutěží</button>
+	</div>
+</div>
+    <div class="row">
+<div class="table-stats-wrapper">
+	<table class="table-stats" id="table--2" style="min-widht: 900px;">
+
+<thead>
+	<tr>
+		<th></th>
+		<th>soutěž</th>
+		<th>klub</th>
+		<th class="hasqtip" data-content="Zápasy" data-hasqtip="22">z</th>
+		<th class="hasqtip" data-content="Góly hráče" data-hasqtip="23">g</th>
+
+		<th class="hasqtip" data-content="Asistence" data-hasqtip="24">a</th>
+		<th class="hasqtip" data-content="Body (G + A)" data-hasqtip="25">b</th>
+		<th class="hasqtip" data-content="Hodnocení hráče za pozitivní a negativní účast na ledě" data-hasqtip="26">+/-</th>
+		<th class="hasqtip" data-content="Hodnocení hráče za pozitivní účast na ledě" data-hasqtip="27">+</th>
+
+		<th class="hasqtip" data-content="Hodnocení hráče za negativní účast na ledě" data-hasqtip="28">-</th>
+		<th class="hasqtip" data-content="Trestné minuty" data-hasqtip="29">tm</th>
+		<th class="hasqtip" data-content="Góly vítězné" data-hasqtip="30">gv</th>
+		<th class="hasqtip" data-content="Góly v přesilovkách" data-hasqtip="31">gp</th>
+
+		<th class="hasqtip" data-content="Góly v oslabení" data-hasqtip="32">go</th>
+		<th></th>
+	</tr>
+</thead>
+		<tbody>
+                    
+					<tr>
+						<th class="table-stats__first" rowspan="3">2014-2015</th>
+
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/34?competition=5579&amp;t=224">
+            Reprezentace 16 (celá sezona)
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/34?competition=5579&amp;teamId=302&amp;t=224">Česko 16</a></td>
+<td class="tbl-c-3">11</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">1</td>
+<td class="tbl-c-6">1</td>
+<td class="tbl-c-7">1</td>
+<td class="tbl-c-8">2</td>
+<td class="tbl-c-9">1</td>
+
+<td class="tbl-c-10">0</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/34?competition=5579&amp;t=224">
+            Reprezentace 16 – příp.
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/34?competition=5579&amp;teamId=302&amp;t=224">Česko 16</a></td>
+<td class="tbl-c-3">8</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">1</td>
+<td class="tbl-c-6">1</td>
+<td class="tbl-c-7">1</td>
+<td class="tbl-c-8">2</td>
+<td class="tbl-c-9">1</td>
+
+<td class="tbl-c-10">0</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/34?competition=5579&amp;t=224">
+            Turnaj čtyř v Rusku – 16
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/34?competition=5579&amp;teamId=302&amp;t=224">Česko 16</a></td>
+<td class="tbl-c-3">3</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">0</td>
+<td class="tbl-c-6">0</td>
+<td class="tbl-c-7">0</td>
+<td class="tbl-c-8">0</td>
+<td class="tbl-c-9">0</td>
+
+<td class="tbl-c-10">0</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+		</tbody>
+		<tbody>
+                    
+					<tr>
+						<th class="table-stats__first" rowspan="6">2015-2016</th>
+
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/33?competition=5619&amp;t=224">
+            Reprezentace 17 (celá sezona)
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/33?competition=5619&amp;teamId=301&amp;t=224">Česko 17</a></td>
+<td class="tbl-c-3">25</td>
+<td class="tbl-c-4">2</td>
+<td class="tbl-c-5">5</td>
+<td class="tbl-c-6">7</td>
+<td class="tbl-c-7">0</td>
+<td class="tbl-c-8">1</td>
+<td class="tbl-c-9">1</td>
+
+<td class="tbl-c-10">10</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/33?competition=5619&amp;t=224">
+            Reprezentace 17 – příp.
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/33?competition=5619&amp;teamId=301&amp;t=224">Česko 17</a></td>
+<td class="tbl-c-3">9</td>
+<td class="tbl-c-4">1</td>
+<td class="tbl-c-5">2</td>
+<td class="tbl-c-6">3</td>
+<td class="tbl-c-7">0</td>
+<td class="tbl-c-8">1</td>
+<td class="tbl-c-9">1</td>
+
+<td class="tbl-c-10">2</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/33?competition=5619&amp;t=224">
+            Turnaj pěti v Arose – 17
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/33?competition=5619&amp;teamId=301&amp;t=224">Česko 17</a></td>
+<td class="tbl-c-3">4</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">1</td>
+<td class="tbl-c-6">1</td>
+<td class="tbl-c-7">0</td>
+<td class="tbl-c-8">0</td>
+<td class="tbl-c-9">0</td>
+
+<td class="tbl-c-10">6</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/33?competition=5619&amp;t=224">
+            World U17 Hockey Challenge
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/33?competition=5619&amp;teamId=301&amp;t=224">Česko 17</a></td>
+<td class="tbl-c-3">5</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">0</td>
+<td class="tbl-c-6">0</td>
+<td class="tbl-c-7">0</td>
+<td class="tbl-c-8">0</td>
+<td class="tbl-c-9">0</td>
+
+<td class="tbl-c-10">0</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/33?competition=5619&amp;t=224">
+            Turnaj čtyř v Södertälje – 17
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/33?competition=5619&amp;teamId=301&amp;t=224">Česko 17</a></td>
+<td class="tbl-c-3">3</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">2</td>
+<td class="tbl-c-6">2</td>
+<td class="tbl-c-7">0</td>
+<td class="tbl-c-8">0</td>
+<td class="tbl-c-9">0</td>
+
+<td class="tbl-c-10">2</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/33?competition=5619&amp;t=224">
+            Turnaj pěti v Plymouthu – 17
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/33?competition=5619&amp;teamId=301&amp;t=224">Česko 17</a></td>
+<td class="tbl-c-3">4</td>
+<td class="tbl-c-4">1</td>
+<td class="tbl-c-5">0</td>
+<td class="tbl-c-6">1</td>
+<td class="tbl-c-7">0</td>
+<td class="tbl-c-8">0</td>
+<td class="tbl-c-9">0</td>
+
+<td class="tbl-c-10">0</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+		</tbody>
+		<tbody>
+                    
+					<tr>
+						<th class="table-stats__first" rowspan="6">2016-2017</th>
+
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/32?competition=5830&amp;t=224">
+            Reprezentace 18 (celá sezona)
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/32?competition=5830&amp;teamId=369&amp;t=224">Česko 18</a></td>
+<td class="tbl-c-3">22</td>
+<td class="tbl-c-4">1</td>
+<td class="tbl-c-5">7</td>
+<td class="tbl-c-6">8</td>
+<td class="tbl-c-7">-1</td>
+<td class="tbl-c-8">2</td>
+<td class="tbl-c-9">3</td>
+
+<td class="tbl-c-10">31</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/32?competition=5830&amp;t=224">
+            Reprezentace 18 – příp.
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/32?competition=5830&amp;teamId=369&amp;t=224">Česko 18</a></td>
+<td class="tbl-c-3">4</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">1</td>
+<td class="tbl-c-6">1</td>
+<td class="tbl-c-7">-1</td>
+<td class="tbl-c-8">0</td>
+<td class="tbl-c-9">1</td>
+
+<td class="tbl-c-10">0</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/32?competition=5830&amp;t=224">
+            Memoriál Ivana Hlinky
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/32?competition=5830&amp;teamId=369&amp;t=224">Česko 18</a></td>
+<td class="tbl-c-3">5</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">0</td>
+<td class="tbl-c-6">0</td>
+<td class="tbl-c-7">0</td>
+<td class="tbl-c-8">2</td>
+<td class="tbl-c-9">2</td>
+
+<td class="tbl-c-10">2</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/32?competition=5830&amp;t=224">
+            Turnaj pěti v Zuchwilu – 18
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/32?competition=5830&amp;teamId=369&amp;t=224">Česko 18</a></td>
+<td class="tbl-c-3">4</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">4</td>
+<td class="tbl-c-6">4</td>
+<td class="tbl-c-7">0</td>
+<td class="tbl-c-8">0</td>
+<td class="tbl-c-9">0</td>
+
+<td class="tbl-c-10">25</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/32?competition=5830&amp;t=224">
+            Turnaj pěti v Plymouthu – 18
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/32?competition=5830&amp;teamId=369&amp;t=224">Česko 18</a></td>
+<td class="tbl-c-3">4</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">1</td>
+<td class="tbl-c-6">1</td>
+<td class="tbl-c-7">0</td>
+<td class="tbl-c-8">0</td>
+<td class="tbl-c-9">0</td>
+
+<td class="tbl-c-10">4</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/32?competition=5830&amp;t=224">
+            MS 18 – 2017
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/32?competition=5830&amp;teamId=369&amp;t=224">Česko 18</a></td>
+<td class="tbl-c-3">5</td>
+<td class="tbl-c-4">1</td>
+<td class="tbl-c-5">1</td>
+<td class="tbl-c-6">2</td>
+<td class="tbl-c-7">0</td>
+<td class="tbl-c-8">0</td>
+<td class="tbl-c-9">0</td>
+
+<td class="tbl-c-10">0</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+		</tbody>
+		<tbody>
+                    
+					<tr>
+						<th class="table-stats__first" rowspan="4">2017-2018</th>
+
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/31?competition=6043&amp;t=224">
+            Reprezentace 20 (celá sezona)
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/31?competition=6043&amp;teamId=140&amp;t=224">Česko 20</a></td>
+<td class="tbl-c-3">16</td>
+<td class="tbl-c-4">1</td>
+<td class="tbl-c-5">0</td>
+<td class="tbl-c-6">1</td>
+<td class="tbl-c-7">-3</td>
+<td class="tbl-c-8">3</td>
+<td class="tbl-c-9">6</td>
+
+<td class="tbl-c-10">4</td>
+<td class="tbl-c-11">1</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/31?competition=6043&amp;t=224">
+            Reprezentace 20 – příp.
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/31?competition=6043&amp;teamId=140&amp;t=224">Česko 20</a></td>
+<td class="tbl-c-3">6</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">0</td>
+<td class="tbl-c-6">0</td>
+<td class="tbl-c-7">2</td>
+<td class="tbl-c-8">2</td>
+<td class="tbl-c-9">0</td>
+
+<td class="tbl-c-10">2</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/31?competition=6043&amp;t=224">
+            Turnaj čtyř ve Finsku – 20
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/31?competition=6043&amp;teamId=140&amp;t=224">Česko 20</a></td>
+<td class="tbl-c-3">3</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">0</td>
+<td class="tbl-c-6">0</td>
+<td class="tbl-c-7">0</td>
+<td class="tbl-c-8">0</td>
+<td class="tbl-c-9">0</td>
+
+<td class="tbl-c-10">0</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/31?competition=6043&amp;t=224">
+            MS 20 – 2018
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/31?competition=6043&amp;teamId=140&amp;t=224">Česko 20</a></td>
+<td class="tbl-c-3">7</td>
+<td class="tbl-c-4">1</td>
+<td class="tbl-c-5">0</td>
+<td class="tbl-c-6">1</td>
+<td class="tbl-c-7">-5</td>
+<td class="tbl-c-8">1</td>
+<td class="tbl-c-9">6</td>
+
+<td class="tbl-c-10">2</td>
+<td class="tbl-c-11">1</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+		</tbody>
+		<tbody>
+                    
+					<tr>
+						<th class="table-stats__first" rowspan="3">2018-2019</th>
+
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/31?competition=6256&amp;t=224">
+            Reprezentace 20 (celá sezona)
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/31?competition=6256&amp;teamId=140&amp;t=224">Česko 20</a></td>
+<td class="tbl-c-3">6</td>
+<td class="tbl-c-4">1</td>
+<td class="tbl-c-5">1</td>
+<td class="tbl-c-6">2</td>
+<td class="tbl-c-7">-3</td>
+<td class="tbl-c-8">1</td>
+<td class="tbl-c-9">4</td>
+
+<td class="tbl-c-10">2</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">1</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/31?competition=6256&amp;t=224">
+            Reprezentace 20 – příp.
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/31?competition=6256&amp;teamId=140&amp;t=224">Česko 20</a></td>
+<td class="tbl-c-3">1</td>
+<td class="tbl-c-4">0</td>
+<td class="tbl-c-5">1</td>
+<td class="tbl-c-6">1</td>
+<td class="tbl-c-7">0</td>
+<td class="tbl-c-8">0</td>
+<td class="tbl-c-9">0</td>
+
+<td class="tbl-c-10">0</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">0</td>
+<td></td>
+					</tr>
+					<tr class="table-stats__highlight table-stats__small">
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/31?competition=6256&amp;t=224">
+            MS 20 – 2019
+		</a>
+</td>
+	<td class="tbl-c-2"><a href="/mladez/player-stats/detailni/31?competition=6256&amp;teamId=140&amp;t=224">Česko 20</a></td>
+<td class="tbl-c-3">5</td>
+<td class="tbl-c-4">1</td>
+<td class="tbl-c-5">0</td>
+<td class="tbl-c-6">1</td>
+<td class="tbl-c-7">-3</td>
+<td class="tbl-c-8">1</td>
+<td class="tbl-c-9">4</td>
+
+<td class="tbl-c-10">2</td>
+<td class="tbl-c-11">0</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">1</td>
+<td></td>
+					</tr>
+		</tbody>
+<tbody>
+	<tr>
+		
+</tr></tbody><thead>
+	<tr>
+		<th></th>
+		<th>soutěž</th>
+		<th></th>
+		<th class="hasqtip" data-content="Zápasy" data-hasqtip="33">z</th>
+		<th class="hasqtip" data-content="Góly hráče" data-hasqtip="34">g</th>
+
+		<th class="hasqtip" data-content="Asistence" data-hasqtip="35">a</th>
+		<th class="hasqtip" data-content="Body (G + A)" data-hasqtip="36">b</th>
+		<th class="hasqtip" data-content="Hodnocení hráče za pozitivní a negativní účast na ledě" data-hasqtip="37">+/-</th>
+		<th class="hasqtip" data-content="Hodnocení hráče za pozitivní účast na ledě" data-hasqtip="38">+</th>
+
+		<th class="hasqtip" data-content="Hodnocení hráče za negativní účast na ledě" data-hasqtip="39">-</th>
+		<th class="hasqtip" data-content="Trestné minuty" data-hasqtip="40">tm</th>
+		<th class="hasqtip" data-content="Góly vítězné" data-hasqtip="41">gv</th>
+		<th class="hasqtip" data-content="Góly v přesilovkách" data-hasqtip="42">gp</th>
+
+		<th class="hasqtip" data-content="Góly v oslabení" data-hasqtip="43">go</th>
+		<th></th>
+	</tr>
+</thead>
+
+		<tbody><tr><th class="table-stats__first" rowspan="1">Vše</th>
+<td class="tbl-c-1">
+		<a href="/mladez/player-stats/detailni/34?t=224">
+            Reprezentační zápasy mládeže
+		</a>
+</td>
+		<td class="tbl-c-2"></td>
+
+<td class="tbl-c-3">80</td>
+<td class="tbl-c-4">5</td>
+<td class="tbl-c-5">14</td>
+<td class="tbl-c-6">19</td>
+<td class="tbl-c-7">-6</td>
+<td class="tbl-c-8">9</td>
+<td class="tbl-c-9">15</td>
+
+<td class="tbl-c-10">47</td>
+<td class="tbl-c-11">1</td>
+<td class="tbl-c-12">0</td>
+<td class="tbl-c-13">1</td>
+<td></td>
+	</tr>
+</tbody>
+	</table>
+</div>
+    </div>
+</div>
+
+		        	<div class="container">
+		        		<div class="banner m-b-10"><!-- %New-on% -->
+<!-- %Zona%:Default;%ID%:1x15;%Banner%:Billboard-bottom-1 -->
+
+<div class="ads"><div class="ads-in" id="Billboard-bottom-1"><script> if(typeof _SAS === 'object') _SAS.ad({site:'Hokej', area:'Ostatni', pos:'Billboard-bottom-1', type:'category'}); </script></div></div>
+</div>
+
+<div class="row p-b-20">
+	<div class="col-45">
+		<div class="banner"><!-- %New-on% -->
+<!-- %Zona%:Default;%ID%:1x13;%Banner%:Fullbanner-left-1 -->
+
+<div class="ads"><div class="ads-in" id="Fullbanner-left-1"><script> if(typeof _SAS === 'object') _SAS.ad({site:'Hokej', area:'Ostatni', pos:'Fullbanner-left-1', type:'category'}); </script></div></div>
+</div>
+	</div>
+	<div class="col-10"></div>
+	<div class="col-45">
+		<div class="banner"><!-- %New-on% -->
+<!-- %Zona%:Default;%ID%:1x14;%Banner%:Fullbanner-right-1 -->
+
+<div class="ads"><div class="ads-in" id="Fullbanner-right-1"><script> if(typeof _SAS === 'object') _SAS.ad({site:'Hokej', area:'Ostatni', pos:'Fullbanner-right-1', type:'category'}); </script></div></div>
+</div>
+	</div>
+</div>
+		        	</div>
+
+		        <div class="footer">
+		        	<div class="container">
+		        		<div class="row">
+		        				<div class="col-25">
+		        					<h2>Hokej.cz</h2>
+<ul class="list">
+	<li><a href="/tipsport-extraliga" title="Tipsport ELH">Tipsport ELH</a></li>
+	<li><a href="/reprezentace" title="Reprezentace">Reprezentace</a></li>
+	<li><a href="/chance-liga" title="CHANCE LIGA">CHANCE LIGA</a></li>
+	<li><a href="/druha-liga" title="2. liga">2. liga</a></li>
+	<li><a href="/ullh" title="ULLH">ULLH</a></li>
+	<li><a href="/mladez" title="Mládež">Mládež</a></li>
+	<li><a href="/chl" title="CHL">CHL</a></li>
+	<li><a href="/zensky-hokej" title="Ženy">Ženy</a></li>
+	<li><a href="/sledge-hokej" title="Para">Para</a></li>
+	<li><a href="/archiv-clanku/rubrika/114" title="Zahraničí">Zahraničí</a></li>
+	<li><a href="http://www.nhl.cz/" title="NHL.cz">NHL.cz</a></li>
+	<li><a href="/tv/hokejka/elh" title="TV">TV</a></li>
+	<li><a href="/hokejka/" title="Hokejka">Hokejka</a></li>
+	<li><a href="/historie" title="Historie">Historie</a></li>
+</ul>		        				</div>
+		        			<div class="col-25">
+		        				<h2>Projekty</h2>
+<ul class="list">
+<li><a href="http://www.hokejistasezony.cz/">Hokejista sezony</a></li>
+<li><a href="http://hokej.cz/zlatahelma">Zlatá helma</a></li>
+<li><a href="http://zlatahokejka.eu">Zlatá hokejka</a></li>
+</ul>
+
+<h2>Organizace</h2>
+
+<ul class="list">
+<li><a href="http://www.bpa.cz">BPA sport marketing</a></li>
+<li><a href="http://www.ceskyhokej.org/">ČSLH</a></li>
+<li><a href="http://www.apklh.cz/">APK LH</a></li>
+<li><a href="http://www.iihf.com">IIHF</a></li>
+<li><a href="http://www.championshockeyleague.net">Champions Hockey League</a></li>
+</ul>
+		        			</div>
+		        				<div class="col-25 footer__social">
+		        					<h2>Sociální sítě</h2>
+<ul class="list-unstyled">
+	<li><a href="https://www.facebook.com/hokejcz" class="btn-facebook"><i class="fa fa-facebook"></i> facebook</a></li>
+	<li><a href="https://twitter.com/hokej_cz" class="btn-twitter"><i class="fa fa-twitter"></i> twitter</a></li>
+	<li><a href="https://www.youtube.com/user/hokejcz" class="btn-youtube"><i class="fa fa-youtube"></i> youtube</a></li>
+</ul>		        				</div>
+<div class="col-25">
+	<a href="https://isport.blesk.cz/ms-hokej-2020" target="_blank" class="isport-link"><img class="isport-logo" src="/images/logo/isport.png" alt="Logo iSport.cz" title="iSport.cz"></a>
+	<ul class="list">
+		<li><a href="https://isport.blesk.cz/clanek/hokej-reprezentace-mistrovstvi-sveta-v-hokeji-2020/364058/program-ms-v-hokeji-2020-kdy-hraji-cesi-ve-skupine-zacnou-se-svedy.html">MS v hokeji</a></li>
+		<li><a href="https://isport.blesk.cz/tipsport-extraliga">TELH</a></li>
+		<li><a href="https://isport.blesk.cz/nhl">NHL</a></li>
+		<li><a href="https://isport.blesk.cz/ms-hokej-u20">MS U20</a></li>
+	</ul>
+</div>
+		        		</div>
+		        	</div>
+		        </div>
+		        <div class="imprint">
+		        	<div class="container">
+		        		<div class="row">
+		        			<div class="col-100 text-center">
+		        				<a href="/rss/vse">RSS</a> | <a href="/kontakt">Kontakt</a> | <a href="/podminky-uziti">Podmínky užití</a> | <a href="/reklama">Reklama</a> -  
+<span title="127000000001">Provozovatel</span> <a href="http://www.bpa.cz/">BPA sport marketing a.s.</a> ve spolupráci s <a href="http://esportsmedia.cz/">eSports.cz, s.r.o.</a>
+
+<p style="margin-top:10px">ISSN 1214-5718 | dotazy na redakci: redakce@hokej.cz, obchod/reklama: obchod@hokej.cz, technický provoz: webmaster@hokej.cz</p>
+
+		        			</div>
+		        		</div>
+		        	</div>
+		        </div>
+		    </div>
+        </div>
+
+        <script>window.jQuery || document.write('<script src="/js/vendor/jquery-1.11.0.min.js"><\/script>')</script><script src="/js/vendor/jquery-1.11.0.min.js"></script>
+<script type="text/javascript" src="/webtemp/jsloader-dc9c3bab55d7.js?1600336098"></script>
+
+		
+		<script type="text/javascript">
+			$(function() {
+
+				$('form.player-search').submit(function(e) {
+					$(this).netteAjax(e);
+					return false;
+				});
+
+				var replacedValue = "--VALUE--";
+				var searchLink = "\/hrac\/23461\/career?playerSearch-q=--VALUE--&stats-section=all&do=playerSearch-search"
+				var searchTimeout;
+
+				$('form.player-search').on('keyup', 'input[type="text"]', function(e) {
+					var $this = $(this);
+					var query = $this.val();
+
+					if (searchTimeout) {
+						clearTimeout(searchTimeout);
+					}
+
+					if (query.length < 2 || !isAllowedKey(e.which)) {
+						return true;
+					}
+
+					searchTimeout = setTimeout(function () {
+						var link = searchLink.replace(replacedValue, $this.val());
+						$.nette.ajax(link);
+					}, 150);
+				}).attr("autocomplete", "off");
+
+				function isAllowedKey(key) {
+					if (key === 0 || key === 13) {
+						return false;
+					}
+
+					return true;
+				}
+			});
+		</script>
+		<script>
+			jQuery.loadScript = function(url, callback) {
+				jQuery.ajax({
+					url:      url,
+					dataType: 'script',
+					success:  callback,
+					async:    true
+				});
+			}
+			function initMlistky(event, club, wac) {
+				$.loadScript("http://"+club+".mlistky.cz/wa1_1.php", function() {
+					new mListky(wac, event, this);
+				});
+			}
+		</script>
+
+		<script type="text/javascript">
+
+        var _gaq = _gaq || [];
+        _gaq.push(['_setAccount', 'UA-180976-5']);
+        _gaq.push(['_trackPageview']);
+
+        _gaq.push(['trac2._setAccount', 'UA-55780611-1']);
+        _gaq.push(['trac2._trackPageview']);
+
+      (function() {
+        var ga = document.createElement('script'); ga.type = 'text/javascript'; ga.async = true;
+        ga.src = ('https:' == document.location.protocol ? 'https://ssl' : 'https://www') + '.google-analytics.com/ga.js';
+         var s = document.getElementsByTagName('script')[0]; s.parentNode.insertBefore(ga, s);
+        })();
+
+</script>
+
+
+<!-- (C)2000-2013 Gemius SA - gemiusAudience / hokej.cz / Homepage -->
+<script type="text/javascript">
+<!--//--><![CDATA[//><!--
+var pp_gemius_identifier = '..swrXBVr9MFfp6D3M_NvbQX3whWp.tN6EoYcGXGzpT.F7';
+// lines below shouldn't be edited
+function gemius_pending(i) { window[i] = window[i] || function() {var x = window[i+'_pdata'] = window[i+'_pdata'] || []; x[x.length]=arguments;};};
+gemius_pending('gemius_hit'); gemius_pending('gemius_event'); gemius_pending('pp_gemius_hit'); gemius_pending('pp_gemius_event');
+(function(d,t) {try {var gt=d.createElement(t),s=d.getElementsByTagName(t)[0],l='http'+((location.protocol=='https:')?'s':''); gt.setAttribute('async','async');
+gt.setAttribute('defer','defer'); gt.src=l+'://spir.hit.gemius.pl/xgemius.js'; s.parentNode.insertBefore(gt,s);} catch (e) {}})(document,'script');
+//--><!]]>
+</script>
+
+
+ 
+<!-- Begin comScore Tag -->
+<script>
+                var _comscore = _comscore || [];
+                _comscore.push({ c1: "2", c2: "9883551" });
+                (function() {
+                               var s = document.createElement("script"), el = document.getElementsByTagName("script")[0]; s.async = true;
+                               s.src = (document.location.protocol == "https:" ? "https://sb" : "http://b") + ".scorecardresearch.com/beacon.js";
+                               el.parentNode.insertBefore(s, el);
+                })();
+</script>
+<noscript>
+                <img src="http://b.scorecardresearch.com/p?c1=2&c2=9883551&cv=2.0&cj=1" />
+</noscript>
+<!-- End comScore Tag -->
+
+<script language="JavaScript" type="text/javascript">
+<!--
+document.write('<img src="https://toplist.cz/dot.asp?id=11657&http='+escape(document.referrer)+'&t='+escape(document.title)+
+'" width="1" height="1" border=0 alt="TOPlist" />'); 
+//--></script><img src="https://toplist.cz/dot.asp?id=11657&amp;http=https%3A//www.hokej.cz/hrac/23461&amp;t=Filip%20Kr%E1l%20%7C%20Hokej.cz%20-%20web%20%u010Desk%E9ho%20hokeje" alt="TOPlist" style="display: none !important;" width="1" hidden="" height="1" border="0">
+
+<script src="https://www.sc.pages06.net/lp/static/js/iMAWebCookie.js?1ddd0b1-156cc89f072-be7239722bf9e6184ab469c11075fd86&amp;h=www.pages06.net" type="text/javascript"></script>
+
+
+<link rel="StyleSheet" href="https://s3-eu-west-1.amazonaws.com/hokej.cz/scripts/ad.css" type="text/css">
+<script src="https://assets.adobedtm.com/4beaca54604aa1db7a7d9296a08d83bee398e7fd/satelliteLib-3ec48de3e3aa4a591bed4da40d172dad01e1032f.js"></script>
+
+<script type="text/javascript">_satellite.pageBottom();</script>
+
+
+
+
+
+    
+
+</body>
+`
