--- utils/date-helpers/get-date-from-array.js
@@ -1,12 +1,0 @@
-module.exports = function (date, y, m, d) {
-  let day = date[d]
-  let month = date[m]
-  const year = date[y]
-
-  month = new Date(Date.parse(`${month} 1, 2012`)).getMonth() + 1
-  month = month < 10 ? `0${month}` : `${month}`
-
-  day = day < 10 ? `0${day}` : `${day}`
-
-  return `${year}-${month}-${day}`
-}
