+++ utils/date/index.js
@@ -1,12 +1,8 @@
-const getCurrentTime = require('./get-current-time')
-const getDateFromArray = require('./get-date-from-array')
 const getMonthName = require('./get-month-name')
 const getSeasonStartYear = require('./get-season-start-year')
 const setDateValues = require('./set-date-values')
 
 module.exports = {
-  getCurrentTime,
-  getDateFromArray,
   getMonthName,
   getSeasonStartYear,
   setDateValues,
